% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_table.R
\docType{data}
\name{annotation_table}
\alias{annotation_table}
\title{Table for the annotation of Barcode}
\format{
A data frame with 5320 rows and 2 columns:
\describe{
  \item{Gene}{It Contains the gene name}
  \item{Barcode}{It contains an ID that identify each barcode, it can be use
  to marge the annotation table with the count table}

  \item{Gene_ID}{It Contains a unique ID of the Gene}

  \item{Sequence}{It contains the cDNA sequence of the Barcode}

  \item{Library}{It contains the library from which the Barcode come from}
}
}
\usage{
data(annotation_table)
}
\description{
Table for the annotation of Barcode
}
\keyword{datasets}
