% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_creation_dataframe.R
\name{barcode_creation_dataframe}
\alias{barcode_creation_dataframe}
\title{Create the table containing the Mapped Sample file}
\usage{
barcode_creation_dataframe(path_file)
}
\arguments{
\item{path_file}{Path of the list of files}
}
\value{
Return the table containing  the list of barcodes mapped in each of
        the samples (a count table).
}
\description{
This function takes as input the path of a list of files
             created using the Samtools idxstats tool and return a
              count table.
}
\concept{compute}
