% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_lost.R
\name{barcode_lost}
\alias{barcode_lost}
\title{Count number of barcode lost}
\usage{
barcode_lost(screenR_Object)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}
}
\value{
Return a tibble containing the number of barcode lost for each
        sample
}
\description{
This function counts the number of barcodes lost during the
             sequencing. A barcode is lost if its associated shRNA has zero
             mapped read in a sample.
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))

# In order to count the number of barcodes lost just the ScreenR object is
# needed
head(barcode_lost(object))

}
\concept{compute}
