% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmpathScreening.R
\name{coef.glmpathScreenr}
\alias{coef.glmpathScreenr}
\title{A method to extract the estimated coefficients from \code{glmpathScreenr} objects.}
\usage{
\method{coef}{glmpathScreenr}(object, intercept = TRUE, or = FALSE)
}
\arguments{
\item{object}{an object of class \code{glmpathScreenr}.}

\item{intercept}{(logical) retain (\code{TRUE}, default) or drop
(\code{FALSE}) the intercept coefficients.}

\item{or}{return odds ratios if \verb{TRUE}; logit-scale coefficients
are the default.}
}
\value{
a \emph{p} x 2 matrix containing the estimated coefficients from the AIC-
and BIC-best logistic regression models, where \emph{p} is the number of coefficients.
}
\description{
A method to extract the estimated coefficients from \code{glmpathScreenr} objects.
}
\examples{
attach(uniobj1)
coef(uniobj)

}
