% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.lasso_screenr}
\alias{coef.lasso_screenr}
\title{A method to extract the estimated coefficients from \code{lasso_screenr} objects.}
\usage{
\method{coef}{lasso_screenr}(object, ..., intercept = TRUE, or = FALSE)
}
\arguments{
\item{object}{an object of class \code{lasso_screenr}.}

\item{...}{optional arguments passed to \code{predict} methods.}

\item{intercept}{(logical) retain (\code{TRUE}, default) or drop
(\code{FALSE}) the intercept coefficients.}

\item{or}{return odds ratios if \verb{TRUE}; logit-scale coefficients
are the default.}
}
\value{
a \emph{p} x 2 matrix containing the estimated coefficients from the AIC-
and BIC-best logistic regression models, where \emph{p} is the number of coefficients.
}
\description{
A method to extract the estimated coefficients from \code{lasso_screenr} objects.
}
\examples{
attach(uniobj1)
coef(uniobj1)
}
