% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticScreening.R
\name{coef.logisticScreenr}
\alias{coef.logisticScreenr}
\title{\code{coef.logisticScreenr} is an S3 coef method for \code{logisticScreenr} objects.}
\usage{
\method{coef}{logisticScreenr}(x, intercept = TRUE, or = FALSE)
}
\arguments{
\item{x}{an object of class \code{logisticScreenr}}

\item{intercept}{(logical) retain (\code{TRUE}, default) or drop
(\code{FALSE}) the intercept coefficients.}

\item{or}{return odds ratios if \verb{TRUE}; logit-scale coefficients
are the default.}
}
\value{
A numeric vector containing the estimated coefficients on the logit
scale.
}
\description{
\code{coef.logisticScreenr} is an S3 coef method for \code{logisticScreenr} objects.
}
\details{
Extracts the estimated coefficients from \code{logisticScreenr} objects.
}
\examples{
load(uniobj2)
class(uniobj2)
coef(uniobj2)

}
