% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.logreg_screenr}
\alias{coef.logreg_screenr}
\title{\code{coef.logreg_screenr} is an S3 coef method for \code{logreg_screenr} objects.}
\usage{
\method{coef}{logreg_screenr}(object, ..., intercept = TRUE, or = FALSE)
}
\arguments{
\item{object}{an object of class \code{logreg_screenr}}

\item{...}{optional arguments passed to \code{predict} methods.}

\item{intercept}{(logical) retain (\code{TRUE}, default) or drop
(\code{FALSE}) the intercept coefficients.}

\item{or}{return odds ratios if \verb{TRUE}; logit-scale coefficients
are the default.}
}
\value{
A numeric vector containing the estimated coefficients on the logit
scale.
}
\description{
\code{coef.logreg_screenr} is an S3 coef method for \code{logreg_screenr} objects.
}
\details{
Extracts the estimated coefficients from \code{logreg_screenr} objects.
}
\examples{
attach(uniobj2)
class(uniobj2)
coef(uniobj2)

}
