% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera_method.R
\name{compute_camera}
\alias{compute_camera}
\title{Compute Camera}
\usage{
compute_camera(
  xglm,
  lrt,
  DGEList,
  matrix_model,
  contrast,
  number_barcode = 3,
  thresh = 1e-04,
  lfc = 1
)
}
\arguments{
\item{xglm}{object created with \code{link{edgeR::estimateDisp}}}

\item{lrt}{object created with \code{link{edgeR::glmFit}}}

\item{DGEList}{object of edgeR}

\item{matrix_model}{the matrix that will be used to perform the
linear model analysis}

\item{contrast}{A vector or a single value indicating the index or the name
of the column the model_matrix wo which perform the analysis}

\item{number_barcode}{Number of barcode to use}

\item{thresh}{The threshold}

\item{lfc}{The Log2FC}
}
\value{
The hit find with the camera method
}
\description{
Compute the actual hit using camera
}
