% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MDS.R
\name{compute_explained_variance}
\alias{compute_explained_variance}
\title{Compute the Explained Variance}
\usage{
compute_explained_variance(screenR_Object)
}
\arguments{
\item{screenR_Object}{The Object of the package}
}
\value{
A data.frame containing all the information of the variance
        expressed by the components
}
\description{
This function compute the explained variance by each of the
             Principal Components
}
\examples{
obj <- get0('obj', envir = asNamespace('ScreenR'))

compute_explained_variance(obj)
}
\keyword{internal}
