% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend.R
\name{compute_trend}
\alias{compute_trend}
\title{Compute Trend}
\usage{
compute_trend(screenR_Object, genes, group_var)
}
\arguments{
\item{screenR_Object}{object created with \code{\link[edgeR]{estimateDisp}}}

\item{genes}{a list of genes}

\item{group_var}{the variable that as to be used as grouping variable}
}
\value{
A table with the trend of the genes passed as input
}
\description{
This function computes the trend of a gene
}
\keyword{internal}
