% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_z_score.R
\name{compute_z_score}
\alias{compute_z_score}
\title{Z-score function}
\usage{
compute_z_score(count_table)
}
\arguments{
\item{count_table}{The count table .}
}
\value{

}
\description{
Create a boxplot to show the distribution of the mapped reads
             in different sample
}
