% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint-methods.R
\name{confint.geeglm}
\alias{confint.geeglm}
\title{An S3 Method to Compute Confidence Limits from \code{geepack::geeglm} Objects.}
\usage{
\method{confint}{geeglm}(object, ..., conf_level = 0.95)
}
\arguments{
\item{object}{an object of class \code{geeglm}.}

\item{...}{optional arguments passed to \code{predict} methods.}

\item{conf_level}{confidence level in (0, 1).}
}
\value{
\code{confint.geeglm} returns a dataframe containing the estimated
model coefficients and their lower and upper confidence limits, \code{lcl}
and \code{ucl}, respectively.
}
\description{
\code{confint.geeglm} returns the logistic model coefficients
estimates and their  normal-theory Wald-type confidence
limits from objects produced by \code{geepack::geeglm}.
}
