% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mapped_reads.R
\name{count_mapped_reads}
\alias{count_mapped_reads}
\title{Count the number of mapped read}
\usage{
count_mapped_reads(screenR_Object)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}
}
\value{
Return a tibble containing the number of mapped read for sample
}
\description{
This function counts the number of reads for each barcode
             in each sample. It is a quality control function (QC) to see if
             the biological protocol went as planned.
             If a sample has very low mapped compared to the other means
             that is has a lower quality.
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
head(count_mapped_reads(object))
}
\concept{compute}
