% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edger_obj.R
\name{create_edger_obj}
\alias{create_edger_obj}
\title{Create EdgeR Object}
\usage{
create_edger_obj(screenR_Object)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}
}
\value{
The edgeR object
}
\description{
Utility function that using the screenR-class
             object create the corresponding edgeR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
create_edger_obj(object)

}
\concept{objects}
