% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Screen_object.R
\name{create_Screen_object}
\alias{create_Screen_object}
\title{create_Screen_object}
\usage{
create_Screen_object(
  count_table,
  annotation_table,
  groups,
  replicates,
  color_palette
)
}
\arguments{
\item{count_table}{The count table obtained from the read alignment that
contains the Barcodes as rows and samples as columns.}

\item{annotation_table}{The annotation table containing the information
for each Barcode and the association to the
corresponding Gene}

\item{groups}{A factor containing the experimental design label}

\item{replicates}{A factor containing the replicates}

\item{color_palette}{a vector containing the color palette}
}
\value{
An object containing all the information for the analysis.
}
\description{
Initial function to create the Screen Object.
}
\details{
Create the Screen Object
}
