% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_zscore_table.R
\name{create_zscore_table}
\alias{create_zscore_table}
\title{Create the Z-score table}
\usage{
create_zscore_table(data_annotated, control, treatment, day)
}
\arguments{
\item{control}{}

\item{treatment}{}

\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}
}
\value{
return a tibble containing the number of barcode lost for sample
}
\description{
This function compute the  Z-score tablestarting from
the screenR object for a given treatment in a given day
}
