% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_mapped_reads.R
\name{distribution_mapped_reads}
\alias{distribution_mapped_reads}
\title{Plot the distribution of the mapped reads using a boxplot}
\usage{
distribution_mapped_reads(
  screenR_Object,
  palette = NULL,
  alpha = 1,
  type = "boxplot"
)
}
\arguments{
\item{palette}{The color vector for the plot.}

\item{alpha}{The opacity of the plot.}

\item{type}{The type of plot. THe default is boxplot the other option is density.}

\item{screenR_object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}.}
}
\value{
return a tibble containing the number of mapped read for sample
}
\description{
Create a boxplot to show the distribution of the mapped reads
             in different sample
}
