% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_tool.R
\name{easy_tool}
\alias{easy_tool}
\title{Simplifying Screening from \code{lasso_screenr}, \code{logreg_screenr} and
\code{gee_screenr} Objects}
\usage{
easy_tool(object, max = 3, model = c("minAIC", "minBIC"), crossval = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{lasso_screenr} or
\code{logreg_screenr}.}

\item{max}{(numeric) the desired maximum value for the response weights.
Default: 3.}

\item{model}{(for \verb{lasso_screenr} objects only) the desired basis
model. Valid options are \verb{"minAIC"} and \verb{"minBIC"}, specifying
the models that produced the smallest AIC and BIC values, respectively.
Default: \verb{minAIC}}

\item{crossval}{a (logical) indicator for cross-validated (\verb{TRUE}) or
in-sample (\verb{FALSE}) performance evaluation. Default: \verb{TRUE}.}

\item{...}{additional arguments passed to \code{coef.lasso_screenr} or
\code{coef.logreg_screenr}.}
}
\value{
\code{easy_tool} returns (invisibly) an object of class \code{easy_tool}
containing:
\describe{
\item{\code{Call}}{The call to \code{easy_tool}.}
\item{\code{varname}}{The names of the response and predictor variables.}
\item{\code{QuestionWeights}}{Weights for the screening questions obtained
by rescaling the non-zero-valued logistic regression coefficients to whole
numbers ranging from 1 to \code{max}.}
\item{\code{Type}}{The type of test performance evaluaion ("cross-validated"
or "in-sample").}
\item{\code{Scores}}{A data frame containing the testing outcomes
(\code{response}) and cross-validated scores obtained as the sums of the
weighted responses to the set of screening questions (\code{score}).}
\item{\code{ROC}}{An object of class \verb{roc} containing the
receiver-operating characteristic produced by \code{`pROC::roc`}.}
}
}
\description{
\code{easy_tool} rescales model coefficients to whole numbers ranging from
1 to \code{max}. Those rescaled and rounded
coefficients can be used as weights (\code{QuestionWeights}) for each
screening question in a simplified model-based screening tool. The test
screening score for a subject is the sum of the weights for their positive
question responses.
}
\details{
The \code{QuestionWeights} (see Value, below) are the foundation for easy
screening. For example, the screening tool could consist of a simple
questionnaire followed by the weight for each question, expressed as a
small whole number (1, ..., \code{max}) and/or an equal number of open
circles.  The person doing the screening need
only circle the numerical weight and/or fill in the circles if and only if the
subject provides a "yes" response to a particular question.  The person doing
the screening then obtains the final score for that subject by adding up the
circled numbers or counting the total number of filled-in circles.  Testing is
mandatory for consenting subjects for whom that final score equals or exceeds
the chosen threshold based on the receiver-operating characteristics of
\code{CVresults}.

The value chosen for \code{max} involves a trade-off between the ease of
manual scoring and the degree to which the ROC from the re-scaling matches the
ROC from the model. Small values of \code{max} make manual scoring easy, and
sufficiently large values will match the screening performance of the model
fit. It is prudent to
compare the ROCs from a few values of \code{max} with the ROC from the model
and base the final choice on the trade-off between ease of manual scoring and
the desired combination of sensitivity and specificity.
}
\examples{
attach(uniobj1)
tool <- easy_tool(uniobj1, max = 3)
methods(class = "easy_tool")
summary(tool)

}
\references{
Teferi W, Gutreuter S, Bekele A et al. Adapting strategies for effective and
efficient pediatric HIV case finding: Risk screening tool for testing children
presenting at high-risk entry points. BMC Infectious Diseases. 2022; 22:480.
\url{http://doi.org/10.1186/s12879-022-07460-w}
}
\seealso{
\code{\link[screenr]{rescale_to_int}}, \code{\link[screenr]{ntpp.easy_tool}},
\code{\link[screenr]{plot.easy_tool}}, \code{\link[screenr]{print.easy_tool}} and
\code{\link[screenr]{summary.easy_tool}}
}
