% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_slope.R
\name{filter_by_slope}
\alias{filter_by_slope}
\title{Filter using the slope filter}
\usage{
filter_by_slope(
  screenR_Object,
  genes,
  group_var_treatment,
  group_var_control,
  slope_control = NULL,
  slope_treatment
)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
create_screenR_object}

\item{genes}{The genes for which the slope as to be computed. Those genes are
the result of the three statistical methods selection}

\item{group_var_treatment}{The variable to use as X for the linear model
for the Treatment}

\item{group_var_control}{The variable to use as X for the linear model
for the the control}

\item{slope_control}{The control slope}

\item{slope_treatment}{The treatment slope}
}
\value{
The screenR_object with the slope for the treatment and the control
        for each gene
}
\description{
This function is used to improve the quality of the hits.
             It computes a regression line in the different samples ad uses
             the slope of this line to see the trend
}
