% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera_method.R
\name{find_camera_hit}
\alias{find_camera_hit}
\title{Find Camera Hit}
\usage{
find_camera_hit(
  screenR_Object,
  matrix_model,
  contrast,
  number_barcode = 3,
  thresh = 1e-04,
  lfc = 1,
  direction = "Down"
)
}
\arguments{
\item{screenR_Object}{The Object of the package}

\item{contrast}{A vector or a single value indicating the index or the name
of the column the model_matrix wo which perform the analysis}

\item{thresh}{The threshold}

\item{lfc}{The Log2FC}

\item{model_matrix}{The matrix that will be used to perform the
linear model analysis}
}
\value{
The hit find with the camera method
}
\description{
Find the hit using the camera method
}
