% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_screenr.R
\name{gee_screenr}
\alias{gee_screenr}
\title{Fitting Screening Tools Using GEE Estimation of Logistic Models}
\usage{
gee_screenr(
  formula,
  id = NULL,
  data = NULL,
  link = c("logit", "cloglog", "probit"),
  corstr = c("independence", "exchangeable", "unstructured"),
  Nfolds = 10,
  partial_auc = c(0.8, 1),
  partial_auc_focus = "sensitivity",
  partial_auc_correct = TRUE,
  boot_n = 4000,
  conf_level = 0.95,
  seed = Sys.time(),
  ...
)
}
\arguments{
\item{formula}{an object of class \code{stats::formula} defining the
testing outcome and predictor covariates, which is passed to \code{stats::glm()}.}

\item{id}{a vector identifying the sampling clusters.}

\item{data}{a dataframe containing the variables defined in \verb{formula}. The
testing outcome must be binary (0,1) indicating negative and positive test
results, respectively, or logical (\verb{TRUE}/\verb{FALSE}).  The covariates
are typically binary (0 = no, 1 = yes) responses to questions which may be
predictive of the test result, but any numeric or factor covariates can be used.}

\item{link}{the character-valued name of the link function for logistic
regression.  Choices are \verb{"logit"}, \verb{"cloglog"} or
\verb{"probit"}. Default: \verb{"logit"}.}

\item{corstr}{a character string specifying the correlation structure. The
following are permitted: \verb{"independence"}, \verb{"exchangeable"} and
\verb{"unstructured"}.  Default: \verb{independence}}

\item{Nfolds}{number of folds used for \emph{k}-fold cross
validation (minimum = 2, maximum = 100).  Default: 10.}

\item{partial_auc}{either a logical \verb{FALSE} or a numeric vector of the
form \code{c(left, right)} where left and right are numbers in the interval
[0, 1] specifying the endpoints for computation of the partial area under the
ROC curve (pAUC). The total AUC is computed if \code{partial_auc} = \verb{FALSE}.
Default: \code{c(0.8, 1.0)}.}

\item{partial_auc_focus}{one of \verb{"sensitivity"} or \verb{specificity},
specifying for which the pAUC should be computed.  \code{partial_auc_focus} is
ignored if \code{partial_auc} = \verb{FALSE}.  Default: \verb{"sensitivity"}.}

\item{partial_auc_correct}{logical value indicating whether the pAUC should be
transformed the interval from 0.5 to 1.0. \code{partial_auc_correct} is
ignored if \code{partial_auc} = \verb{FALSE}. Default: \verb{TRUE}).}

\item{boot_n}{Number of bootstrap replications for computation of confidence
intervals for the (partial)AUC. Default: 4000.}

\item{conf_level}{a number between 0 and 1 specifying the confidence level
for confidence intervals for the (partial)AUC. Default: 0.95.}

\item{seed}{random-number generator seed for cross-validation data splitting.}

\item{...}{additional arguments passsed to or from other \code{geepack::geeglm}
or \code{pROC::roc}.}
}
\value{
\code{gee_screenr} returns an object of class \code{gee_screenr},
which inherits from class \code{logreg_screenr}, containing the elements:
\describe{
\item{\code{Call}}{The function call.}
\item{\code{formula}}{The formula object.}
\item{\code{Prevalence}}{Prevalence (proportion) of the test condition in the
training sample.}
\item{\code{ModelFit}}{An object of class \verb{glm} (See \code{\link{glm}})
containing the results of the model fit.}
\item{\code{ISroc}}{An object of class \code{\link[pROC]{roc}} containing
the "in-sample" (overly-optimistic) receiver operating characteristics,
and additional functions for use with this object are available in the
\code{pROC} package.}
\item{\code{CVpreds}}{An object of class \code{cv.predictions} containing
the data and cross-validated predicted condition \code{y}.}
\item{\code{CVroc}}{An object of class \code{\link[pROC]{roc}} containing
the \emph{k}-fold cross-validated "out-of-sample" receiver operating
characteristics, and additional functions for use with this object are
available in the \code{pROC} package.}
\item{\code{CVcoef}}{the estimated coefficients from cross-validation}
\item{\code{X_ho}}{{the matrix of held-out predictors for each cross-validation fold}}
}
}
\description{
\code{gee_screenr} is a convenience function which integrates GEE estimation
of logsitic models, \emph{k}-fold  cross-validation and estimation of the
receiver-operating characteristic. GEE estimation accommodates cluster sampling.
}
\details{
The results provide information from
which to choose a probability threshold above which individual out-of-sample
probabilies indicate the need to perform a diagnostic test.  Out-of-sample
performance is estimated using \emph{k}-fold cross validation.

The receiver operating characteristics are computed using the \code{pROC}
package. See References and package documentation for additional details.

By default, the \emph{partial} area under the ROC curve is computed from
that portion of the curve for which sensitivity is in the closed interval
[0.8, 1.0]. However, the total AUC can be obtained using the argument
\code{partial_auc = FALSE}.  Partial areas can be computed for either
ranges of sensitivity or specificity using the arguments
\code{partial_auc_focus} and \code{partial_auc}.  By default, partial areas
are standardized.

Out-of-sample performance is estimated using \emph{k}-fold cross-validation.
For a gentle but python-centric introduction to \emph{k}-fold cross-validation,
see \url{https://machinelearningmastery.com/k-fold-cross-validation/}.
}
\examples{
\dontrun{
library(dplyr)
data(unicorns)
## Add a contrived cluster identifier (25 clusters) for demonstration only:
uniclus <- unicorns \%>\%
   mutate(cluster = sample(1:25, size = dim(unicorns)[1], replace = TRUE))
## Use gee_screenr:
uniobj3 <- gee_screenr(testresult ~ Q1 + Q2 + Q3 + Q5 + Q6 + Q7, id = cluster,
                       data = uniclus, link = "logit", Nfolds = 10)
class(uniobj3)
methods(class = class(uniobj3)[1])
methods(class = class(uniobj3)[2])
summary(uniobj3)
}
}
\references{
Liang K-Y, Zeger SL. Longitudinal data analysis using generalized linear
models. Biometrika 1986;73(1):13-22.
\url{http://doi.org/10.2307/2336267}

Halekoh U, Hojsgaard S, Yan, J. The R package geepack for generalized
estimating equations. Journal of Statistical Software 2006;15(2):1-11.
\url{http://doi.org/10.18637/jss.v015.i02}

Kim J-H. Estimating classification error rate: Repeated cross-validation, repeated
hold-out and bootstrap. Computational Statistics and Data Analysis
2009:53(11):3735-3745. \url{http://doi.org/10.1016/j.csda.2009.04.009}

Robin X, Turck N, Hainard A, Tiberti N, Lisacek F, Sanchez J-C,
Muller M. \code{pROC}: An open-source package for \code{R} and S+ to
analyze and compare ROC curves. BMC Bioinformatics 2011;12(77):1-8.
\url{http://doi.org/10.1186/1471-2105-12-77}
}
\seealso{
\code{\link[geepack]{geeglm}}, \code{\link[pROC]{roc}} and
\code{\link[pROC]{auc}}
}
