% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenr-class.R
\name{get_annotation_table,screenr_object-method}
\alias{get_annotation_table,screenr_object-method}
\alias{get_annotation_table,screenr_object}
\title{Get ScreenR annotation table}
\usage{
\S4method{get_annotation_table}{screenr_object}(object)
}
\arguments{
\item{object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}
}
\value{
The annotation table of the ScreenR object
}
\description{
Get function for the annotation table of the ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
annotation_table <- get_annotation_table(object)
head(annotation_table)
}
\concept{objects}
