% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenr-class.R
\name{get_count_table}
\alias{get_count_table}
\title{Get ScreenR count table}
\usage{
get_count_table(object)
}
\arguments{
\item{object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}
}
\value{
The count table of the ScreenR object
}
\description{
Get function for the count table of the ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
count_table <- get_count_table(object)
head(count_table)
}
\concept{objects}
