% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/screenr-class.R
\name{get_data_table}
\alias{get_data_table}
\alias{get_data_table,screenr_object-method}
\alias{get_data_table,screenr_object}
\title{Get ScreenR data_table}
\usage{
get_data_table(object)

\S4method{get_data_table}{screenr_object}(object)
}
\arguments{
\item{object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}
}
\value{
The data_table of the ScreenR object
}
\description{
Get function for the data_table of the ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
data_table <- get_data_table(object)
data_table
}
\concept{objects}
