% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenr-class.R
\name{get_groups,screenr_object-method}
\alias{get_groups,screenr_object-method}
\alias{get_groups,screenr_object}
\title{Get ScreenR groups}
\usage{
\S4method{get_groups}{screenr_object}(object)
}
\arguments{
\item{object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}
}
\value{
The groups of the ScreenR object
}
\description{
Get function for the groups of the ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
groups <- get_groups(object)
groups
}
\concept{objects}
