% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenr-class.R
\name{get_normalized_count_table,screenr_object-method}
\alias{get_normalized_count_table,screenr_object-method}
\alias{get_normalized_count_table,screenr_object}
\title{Get ScreenR normalized_count_table}
\usage{
\S4method{get_normalized_count_table}{screenr_object}(object)
}
\arguments{
\item{object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}
}
\value{
The normalized_count_table of the ScreenR object
}
\description{
Get function for the normalized_count_table of
             the ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
normalized_count_table <- get_normalized_count_table(object)
normalized_count_table
}
\concept{objects}
