% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/screenr-class.R
\name{get_replicates}
\alias{get_replicates}
\alias{get_replicates,screenr_object-method}
\alias{get_replicates,screenr_object}
\title{Get ScreenR replicates}
\usage{
get_replicates(object)

\S4method{get_replicates}{screenr_object}(object)
}
\arguments{
\item{object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}
}
\value{
The replicates of the ScreenR object
}
\description{
Get function for the replicates of the ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
replicates <- get_replicates(object)
replicates
}
\concept{objects}
