% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_what-methods.R
\name{get_what.lasso_screenr}
\alias{get_what.lasso_screenr}
\title{An S3 Method for Extraction of Components from \code{lasso_screenr} Objects}
\usage{
\method{get_what}{lasso_screenr}(
  from = NULL,
  what = c("glmpathObj", "ROCci", "cvROC", "isROC"),
  ...,
  model = c("minAIC", "minBIC"),
  conf_level = 0.95,
  bootreps = 4000,
  se_min = 0.8
)
}
\arguments{
\item{from}{the \code{lasso_screenr}-class object from which to extract
the component.}

\item{what}{the character-valued name of the component to extract. Valid values are
\verb{"glmpathObj"}, \verb{"ROCci"}, \verb{"cvROC"} and \verb{"isROC"}.}

\item{...}{optional arguments to \code{get_what} methods.}

\item{model}{the character-valued name of the model for which the component is
desired.  Valid values are \verb{"minAIC"} and \verb{"minBIC"}.
Default: \verb{"minAIC"}.}

\item{conf_level}{confidence level for \code{what =} \verb{"ROCci"}. Default: 0.95.}

\item{bootreps}{the number of bootstrap replications for estimation of
confidence intervals for \code{what =} \verb{"ROCci"}. Default: 4000.}

\item{se_min}{minimum value of sensitivity printed for
\code{what =} \verb{ROCci}.  Default: 0.8.}
}
\value{
\code{get_what.lasso_screenr} returns (invisibly) the object specified
by \code{what}.
}
\description{
\code{get_what.lasso_screenr} extracts components from \code{lasso_screenr}-class
objects.
}
\details{
\code{get_what} is provided to enable easy extraction of components that are
not provided by the \code{coef}, \code{plot}, \code{predict}, \code{print}
or \code{summary} methods.

The following values of \code{what} return:
\describe{
\item{\verb{"glmpathObj"}}{the entire \code{glmpath}-class object produced by
by \code{\link[glmpath]{glmpath}}}.
\item{\verb{ROCci}}{a data frame containing cross-validated sensitivities,
specificities and their confidence limits, and thresholds}.
\item{\verb{"cvROC"}}{the \code{roc}-class object produced by \code{\link[pROC]{roc}}
containing the \emph{k}-fold cross-validated receiver-operating characteristic.}
\item{\verb{"isROC"}}{the \code{roc}-class object produced by \code{\link[pROC]{roc}}
containing the in-sample (overly optimistic) receiver-operating characteristic.}
}
}
\examples{
\dontrun{
attach(uniobj1)
## Plot the coefficient paths
pathobj <- get_what(from = uniobj1, what = "glmpathObj", model = "minAIC")
plot(pathobj, xvar = "lambda")
## Get and print cross-validated sensitivities and specificities at
##   thresholds for the local maxima of the ROC curve
cvROCci <- get_what(from = uniobj1,  what = "ROCci", model = "minBIC")
print(cvROCci)
}

}
