% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_what-methods.R
\name{get_what}
\alias{get_what}
\title{S3 Methods for Extraction of Object Components}
\usage{
get_what(from, what, ...)
}
\arguments{
\item{from}{an object from which to extract \code{what}.}

\item{what}{the element to extract from \code{from}.}

\item{...}{additional arguments.}
}
\value{
\code{get_what} returns the object specified by \code{what}.
}
\description{
\code{get_what} extracts components from objects.
}
\seealso{
\code{\link[screenr]{get_what.easy_tool}},
\code{\link[screenr]{get_what.lasso_screenr}},
\code{\link[screenr]{get_what.logreg_screenr}} and
\code{\link[screenr]{get_what.simple_screenr}}.
}
