% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_what.R
\name{get_what.simple_screenr}
\alias{get_what.simple_screenr}
\title{\code{get_what.simple_screenr} is an S3 method to extract components of
\code{simple_screenr} objects.}
\usage{
\method{get_what}{simple_screenr}(from = NULL, what = NULL, ..., conf_level = 0.95, bootreps = 2000)
}
\arguments{
\item{from}{the \code{simple_screenr}-class object from which to extract
the component.}

\item{what}{the (character) name of the component to extract. Valid values
are \verb{"ROCci"} and \verb{"isROC"}.}

\item{...}{optional arguments to \code{get_what} methods.}

\item{conf_level}{(optional) confidence level for \code{what =} \verb{"ROCci"}}

\item{bootreps}{the number of bootstrap replications for estimation of
confidence intervals for \code{what =} \verb{"ROCci"}.}
}
\value{
The selected component is returned invisibly.
}
\description{
\code{get_what.simple_screenr} is an S3 method to extract components of
\code{simple_screenr} objects.
}
\details{
\code{get_what} is provided to enable easy extraction of components for those
who wish to perform computations that are not provided by the
\code{plot}, \code{predict}, \code{print} or \code{summary} methods.

The following values of \code{what} return:
\describe{
\item{\verb{"isROC"}}{the \code{roc}-class object produced by \code{\link[pROC]{roc}}
containing the in-sample (overly optimistic) receiver-operating characteristic.}
}
}
\examples{
\dontrun{
data(unicorns)
too_simple <- simple_screenr(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5 + Q6 + Q7,
                            data = unicorns)
roc <- get_what(from = too_simple, what = "isROC" )
plot(roc)
}

}
