% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialScreening.R
\name{getCoefficients}
\alias{getCoefficients}
\title{Extract the estimated coefficients from \code{binomscreenr} objects}
\usage{
getCoefficients(x)
}
\arguments{
\item{x}{An object of class \code{biomscreenr}}
}
\value{
A numeric vector containing the estimated coefficients on the logit
scale.
}
\description{
Extract the estimated coefficients from \code{binomscreenr} objects
}
