% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialScreening.R
\name{getOddsRatios}
\alias{getOddsRatios}
\title{Extract a dataframe containing odds ratios and their profile-likelihood confidence intervals}
\usage{
getOddsRatios(x, level = 0.95)
}
\arguments{
\item{x}{an object of class \code{biomscreenr}}

\item{level}{the desired confidence level.  The default is 0.95.}
}
\value{
A dataframe containing:
\describe{
\item{\code{Variable}}{The names of predictor covariates.}
\item{\code{OddsRatio}}{The estimated odds ratios.}
\item{\code{LCL}}{Lower confidence limit}
\item{\code{UCL}}{Uppper confidence limit}
}
}
\description{
Extract a dataframe containing odds ratios and their profile-likelihood confidence intervals
}
