% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyTool.R
\name{getWhat.easyTool}
\alias{getWhat.easyTool}
\title{\code{getWhat.easyTool} is an S3 method to extract components of
\code{easyTool} objects.}
\usage{
\method{getWhat}{easyTool}(from = NULL, what = NULL, model = NULL)
}
\arguments{
\item{from}{the \code{easyTool}-class object from which to extract
the component.}

\item{what}{the (character) name of the component to extract. Valid values are
\verb{"Call"}, \verb{"QuestionWeights"}, \verb{"ROC"} and \verb{"Scores"}.}
}
\value{
The selected component is returned invisibly.
}
\description{
\code{getWhat.easyTool} is an S3 method to extract components of
\code{easyTool} objects.
}
\details{
\code{getWhat} is provided to enable easy extraction of components for those
who wish to perform computations that are not provided by the
\code{plot}, \code{predict}, \code{print} or \code{summary} methods.

The following values of \code{what} return:
\describe{
\item{\verb{"Call"}}{the function call that created \code{from}.}
\item{\verb{"QuestionWeights"}}{the screening question weights, which are the re-scale
logistic-regression coefficients.}
\item{\verb{"Scores"}}{the screening scores for each subject, which are the sums of the
products of the binary question responses and their \code{QuestionWeights}}
\item{\verb{"ROC"}}{the receiver-operating characteristic for the \code{Scores}}
}
}
\examples{
attach(uniobj1)
tool <- easyTool(uniobj1, max = c, crossval = TRUE)
weights <- getWhat(from = tool, what = "QuestionWeights")
print(weights)

}
