% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmpathScreening.R
\name{getWhat.glmpathScreenr}
\alias{getWhat.glmpathScreenr}
\title{\code{getWhat.glmpathScreenr} is an S3 method to extract components of
\code{glmpathScreenr} objects.}
\usage{
\method{getWhat}{glmpathScreenr}(from = NULL, what = NULL, model = NULL)
}
\arguments{
\item{from}{the \code{glmpathScreenr}-class object from which to extract
the component.}

\item{what}{the (character) name of the component to extract. Valid values are
\verb{"glmpathObj"}, \verb{"cvROC"} and \verb{"isROC"}.}

\item{model}{the (character) name of the model for which the component is
desired.  Valid values are \verb{"minAIC"} and \verb{"minBIC"}.}
}
\value{
The selected component is returned invisibly.
}
\description{
\code{getWhat.glmpathScreenr} is an S3 method to extract components of
\code{glmpathScreenr} objects.
}
\details{
\code{getWhat} is provided to enable easy extraction of components for those
who wish to perform computations that are not provided by the \code{coef},
\code{plot}, \code{predict}, \code{print} or \code{summary} methods.

The following values of \code{what} return:
\describe{
\item{\verb{"glmpathObj"}}{the entire \code{glmpath}-class object produced by
by \code{\link[glmpath]{glmpath}}}.
\item{\verb{"cvROC"}}{the \code{roc}-class object produced by \code{\link[pROC]{roc}}
containing the \emph{k}-fold cross-validated receiver-operating characteristic.}
\item{\verb{"isROC"}}{the \code{roc}-class object produced by \code{\link[pROC]{roc}}
containing the in-sample (overly optimistic) receiver-operating characteristic.}
}
}
\examples{
attach(uniobj1)
pathobj <- getWhat(from = uniobj1, what = "glmpathObj", model = "minAIC")
plot(pathobj)
cvROC <- getWhat(from = uniobj1,  what = "cvROC", model = "minBIC")
plot(cvROC)

}
