% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticScreening.R
\name{getWhat.logisticScreenr}
\alias{getWhat.logisticScreenr}
\title{\code{getWhat.logisticScreenr} is an S3 method to extract components of
\code{logisticScreenr} objects.}
\usage{
\method{getWhat}{logisticScreenr}(from = NULL, what = NULL)
}
\arguments{
\item{from}{the \code{logisticScreenr}-class object from which to extract
the component.}

\item{what}{the (character) name of the component to extract. Valid values are
\verb{"ModelFit"}, \verb{"cvROC"} and \verb{"isROC"}.}
}
\value{
The selected component is returned invisibly.
}
\description{
\code{getWhat.logisticScreenr} is an S3 method to extract components of
\code{logisticScreenr} objects.
}
\details{
\code{getWhat} is provided to enable easy extraction of components for those
who wish to perform computations that are not provided by the \code{coef},
\code{plot}, \code{predict}, \code{print} or \code{summary} methods.

The following values of \code{what} return:
\describe{
\item{\verb{"ModelFit"}}{the entire \code{glm}-class object produced by
by \code{\link[stats]{glm}}}.
\item{\verb{"cvROC"}}{the \code{roc}-class object produced by \code{\link[pROC]{roc}}
containing the \emph{k}-fold cross-validated receiver-operating characteristic.}
\item{\verb{"isROC"}}{the \code{roc}-class object produced by \code{\link[pROC]{roc}}
containing the in-sample (overly optimistic) receiver-operating characteristic.}
}
}
\examples{
load(uniobj2)
class(uniobj2)
mfit <- getWhat(from = uniobj2, what = "ModelFit")
mfit$coefs

}
