% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getWhat}
\alias{getWhat}
\title{\code{getWhat} is an S3 generic function to extract components of objects
produced by functions in the \code{screenr} package}
\usage{
\code{getWhat(object, ...)}
}
\description{
\code{getWhat} is an S3 generic function to extract components of objects
produced by functions in the \code{screenr} package
}
\seealso{
\code{link[screenr]{getWhat.glmpathScreenr}}

\code{link[screenr]{getWhat.logisticScreenr}}
}
