% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWhat.R
\name{getWhat.simpleScreenr}
\alias{getWhat.simpleScreenr}
\title{\code{getWhat.simpleScreenr} is an S3 method to extract components of
\code{simpleScreenr} objects.}
\usage{
\method{getWhat}{simpleScreenr}(from = NULL, what = "isROC", ...)
}
\arguments{
\item{from}{the \code{simpleScreenr}-class object from which to extract
the component.}

\item{what}{the (optional character) name of the component to extract. The
only valid value is \verb{"isROC"}.}

\item{...}{optional arguments to \code{getWhat} methods.}
}
\value{
The selected component is returned invisibly.
}
\description{
\code{getWhat.simpleScreenr} is an S3 method to extract components of
\code{simpleScreenr} objects.
}
\details{
\code{getWhat} is provided to enable easy extraction of components for those
who wish to perform computations that are not provided by the
\code{plot}, \code{predict}, \code{print} or \code{summary} methods.

The following values of \code{what} return:
\describe{
\item{\verb{"isROC"}}{the \code{roc}-class object produced by \code{\link[pROC]{roc}}
containing the in-sample (overly optimistic) receiver-operating characteristic.}
}
}
\examples{
\dontrun{
data(unicorns)
toosimple <- simpleScreenr(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5 + Q6,
                           data = unicorns)
roc <- getWhat(from = toosimple, what = "isROC" )
plot(roc)
}

}
