% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmpathScreening.R
\name{glmpathScreenr}
\alias{glmpathScreenr}
\title{\code{glmpathScreenr} is a used to develop test-screening tools based on \emph{L}1
regularization of logistic regression based on \code{\link[glmpath]{glmpath}}.}
\usage{
glmpathScreenr(formula, data = NULL, Nfolds = 10, seed = Sys.time(), ...)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}} defining the
testing outcome and predictor covariates}

\item{data}{a dataframe containing the variables defined in \verb{formula}}

\item{Nfolds}{the number of folds used for \emph{k}-fold cross
validation (default = 10, minimum = 2, maximum = 100).}

\item{seed}{RNG seed for cross-validation data splitting}

\item{...}{additional arguments passed to glmpath::glmpath or pROC::roc}

\item{criterion}{information criterion (IC) used to select the best model
(\verb{"AIC"} or \verb{"BIC"})}
}
\value{
Return (invisibly) an object of class \code{glmpathScreenr} containing
the elements:
\describe{
\item{\code{Call}}{The function call.}
\item{\code{Prevalence}}{Prevalence of the binary response variable.}
\item{\code{glmpathObj}}{An object of class \code{glmpath} returned by
\code{glmpath::glmpath}. See \code{help(glmpath)} and
\code{methods(class = "glmpath")}.}
\item{\code{Xmat}}{The matrix of predictors.}
\item{\code{isResults}}{A list structure containing the results from the two
model fits which produced the minimum AIC and BIC values, respectively. The
results consist of \code{Coefficients} (the logit-scale parameter estimates,
including the intercept), \code{isPreds} (the in-sample predicted
probabilities) and \code{isROC} (the in-sample receiver-operating
characteristic (ROC) of class \code{roc}).}
\item{\code{cvResults}}{A list structure containing the results of \emph{k}-
fold cross-validation estimation of out-of-sample performance.  The list
elements are:}
\describe{
\item{\code{Nfolds}}{the number folds \emph{k}}
\item{\code{X_ho}}{the matrix of held-out predictors for each cross-validation fold}
\item{\code{minAICcvPreds}}{the held-out responses and out-of-sample predicted probabilities from
AIC-best model selection}
\item{\code{minAICcvROC}}{the out-of-sample ROC object
of class \code{roc} from AIC-best model selection}
\item{\code{minBICcvPreds}}{the held-out responses and out-of-sample predicted probabilities from
BIC-best model selection}
\item{\code{minBICcvROC}}{the corresponding out-of-sample predicted probabilities
and ROC object from BIC-best model selection}
}
\item{\code{RNG}}{Specification of the random-number generator used for
k-fold data splitting.}
\item{\code{RNGseed}}{RNG seed.}
}
}
\description{
\code{glmpathScreenr} is a used to develop test-screening tools based on \emph{L}1
regularization of logistic regression based on \code{\link[glmpath]{glmpath}}.
}
\details{
\code{glmpathScreenr} is a convenience function which integrates
logistic regression using GLM-path \emph{L}1 regularization, \emph{k}-fold
cross-validation and estimation of the receiver-operating characteristic.
The in-sample and out-of-sample performance is estimated from the models
which produced the minimum AIC and minimum BIC.  Execute
\code{methods(class = "glmpathScreenr")} to identify available methods.

The \emph{L}1 path regularizer of Park and Hastie (2007) is similar
to the more familiar lasso and elastic net. It differs from the
lasso with the inclusion of a small fixed \emph{L}2 penalty, and
differs from the elastic net in that the \emph{L}2 penalty is
fixed.  Like the elastic net, the Park-Hastie \emph{L}1 path
regularization is robust to highly correlated predictors.

For a gentle but Python-centric introduction to \emph{k}-fold cross-validation,
see \link{https://machinelearningmastery.com/k-fold-cross-validation/}.
}
\examples{
data(unicorns)
help(unicorns)
uniobj1 <- glmpathScreenr(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5 + Q6,
                            data = unicorns, Nfolds = 10)
summary(uniobj1)

}
\references{
Park MY, Hastie T. \emph{L}1-regularization path algorithm for generalized linear
models. Journal of the Royal Statistical Society Series B. 2007;69(4):659-677.
\url{https://doi.org/10.1111/j.1467-9868.2007.00607.x}

Robin X, Turck N, Hainard A, Tiberti N, Lisacek F, Sanchez J-C,
Müller M. \code{pROC}: An open-source package for \code{R} and S+ to
analyze and compare ROC curves. BMC Bioinformatics. 2011;12(77):1-8.
\url{http://doi.org/10.1186/1471-2105-12-77}
}
\seealso{
\code{\link[glmpath]{glmpath}}, \code{link[pROC]{roc}}
}
