% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{inverseLink}
\alias{inverseLink}
\title{Compute Inverses of Binomial Regression Link Functions}
\usage{
inverseLink(link, lp)
}
\arguments{
\item{link}{Character link function (one of \verb{"logit"}, \verb{"cloglog"}
or \verb{"probit"})}

\item{lp}{Numeric vector containin the linear predictor}
}
\value{
A numeric vector containing the inverse of the link function for the
linear predictor.
}
\description{
Returns the inverse of logit, cloglog and probit link functions for a linear
predictor
}
