% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{keepfirst}
\alias{keepfirst}
\title{\code{keepfirst} returns only those rows of a dataframe having unique values
in selected columns.}
\usage{
keepfirst(x, colnames, data = NULL)
}
\arguments{
\item{x}{character-valued column name along which the dataframe is sorted.}

\item{colnames}{a character vector of column names  to identify uniqueness.}

\item{data}{a data frame.}
}
\value{
A data frame consisting of the rows of \code{data} which are
unique with respect to \code{colnames}
}
\description{
\code{keepfirst} returns only those rows of a dataframe having unique values
in selected columns.
}
\details{
The dataframe \code{data} is sorted, and then only those rows which are unique
with respect to the values of selected columns.
}
