% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapped_reads.R
\name{mapped_reads}
\alias{mapped_reads}
\title{Mapped Reads}
\usage{
mapped_reads(screenR_Object)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}
}
\value{
Return a tibble containing the number of mapped read for sample
}
\description{
This function returns the number of mapped reads inside the
             ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
mapped_reads(object)

}
\concept{compute}
