% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{nnt_}
\alias{nnt_}
\title{Compute the Ratio of Total Tests Performed  Per Postive Result}
\usage{
nnt_(dframe)
}
\arguments{
\item{dframe}{a dataframe containing columns \code{sensitivities},
\code{specificities} and \code{prev}.}
}
\value{
\code{nnt_} returns adataframe containing sensitivity,
specificity, the anticipated
average number of tests required to observe a single positive test
result \code{ntpp}, and the prevalence among those screened out
of testing \code{pre_untested}.
}
\description{
\code{nnt_} computes the anticipated average number of
tests performed in order to observe a positive test result.
}
