% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize data}
\usage{
normalize_data(screenR_Object)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}
}
\value{
Return the ScreenR object with the normalize data
}
\description{
This function perform a normalization on the data considering
             the fact that each shRNA has a defined length so this will not
             influence the data. Basically is computed the sum for
             each row and then multiply by 1e6. At the end we will obtain a
             sort of transcript per million.
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
object <- normalize_data(object)

slot(object, "normalized_count_table")
}
\concept{compute}
