% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{nSensSpec}
\alias{nSensSpec}
\title{Sample Size for Joint Testing of Sensitivity and Specificity}
\usage{
nSensSpec(Sens, Spec, SnsCrit = 0.9, SpcCrit = 0.9, alpha = 0.05,
  power = 0.8)
}
\arguments{
\item{Sens}{A vector of anticipated sensitivities (0 < Sens < 1).}

\item{Spec}{A vector of anticipated specificities (0 < Spec < 1).}

\item{SnsCrit}{Critical value for sensitivity.}

\item{SpcCrit}{Critical value for specificity.}

\item{alpha}{Probability of Type I error.}

\item{power}{Desired power.}
}
\value{
A data frame containing the required number of positive and negative
test results.
}
\description{
Estimate the required number of positive and negative test results for
hypothesis testing against the joint null hypothesis H0: Sens = SnsCrit and
Spec = SpcCrit.
}
\examples{
sens <- c(rep(0.8, 4), rep(0.85, 4), rep(0.90, 4))
spec <- rep(c(0.65, 0.70, 0.75, 0.8 ), 3)
nSensSpec(sens, spec, SnsCrit = 0.90, SpcCrit = 0.70)
}
\references{
Sullivan, M.P., The Statistical Evaluation of Medical Tests
for Classification and Prediction.  Oxford Statistical Science Series 28.
Oxford University Press, Oxford UK.
}
