% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{nSensSpec_est}
\alias{nSensSpec_est}
\title{Sample Size for Joint Estimation of Sensitivity and Specificity}
\usage{
nSensSpec_est(Sens, Spec, w = 0.1, CL = 0.95)
}
\arguments{
\item{Sens}{A numeric vector of anticipated sensitivities (0 < Sens < 1).}

\item{Spec}{A numeric vector of anticipated specificities (0 < Spec < 1).}

\item{w}{confidence interval margin of error (half width)}

\item{CL}{Confidence level for joint rectangular confidence intervals.}
}
\value{
A data frame containing the required number of positive and negative
test results.
}
\description{
Estimate the required number of positive and negative test results for
estimation of joint rectangular confidence intervals for sensitivity and
specificity.  For rare conditions, the number of positive test results
will be limiting, and sample size should be based on that alone.
}
\examples{
sens <- c(rep(0.8, 4), rep(0.85, 4), rep(0.90, 4))
spec <- rep(c(0.65, 0.70, 0.75, 0.8 ), 3)
nSensSpec_est(sens, spec, w = 0.1)
}
\references{
Pepe, M.S. The Statistical Evaluation of Medical Tests for
Classification and Prediction.  Oxford University Press, 2003. Oxford UK.
}
