% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{ntpp.data.frame}
\alias{ntpp.data.frame}
\title{\code{ntpp.data.frame} is a method for computation of the anticipated
number of tests per positive test result}
\usage{
\method{ntpp}{data.frame}(dframe)
}
\arguments{
\item{dframe}{a one-row dataframe containing columns \code{sensitivity},
\code{specificity} and \code{prev}.}
}
\value{
a data frame containing the following columns:
\describe{
\item{\code{sensitivity}}{the sensitivity (proportion)}
\item{\code{specificity}}{the specificity (proportion)}
\item{\code{prev}}{prevalence proportion of the test condition}
\item{\code{ntpp}}{anticipated total tests required per positive result}
\item{\code{prev_untested}}{anticipated prevalence proportion among the untested}
}
}
\description{
\code{ntpp.data.frame} is a method for computation of the anticipated
number of tests per positive test result
}
