% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntpp-methods.R
\name{ntpp.default}
\alias{ntpp.default}
\title{An S3 Method to Compute the Ratio of Total Tests to Positive Results}
\usage{
\method{ntpp}{default}(object = NULL, ..., se = NULL, sp = NULL, prev = NULL)
}
\arguments{
\item{object}{unused, specify \code{se}, \code{sp} and \code{prev}}

\item{...}{optional arguments to \code{ntpp} methods.}

\item{se}{a numeric vector of sensitivities in (0,1)}

\item{sp}{a numeric vector of sensitivities in (0,1)}

\item{prev}{a numeric vector of prevalences of the testing condition, in (0,1)}
}
\value{
\code{ntpp.default} returns a data frame containing the
following columns:
\describe{
\item{\code{sensitivity}}{the sensitivity (proportion)}
\item{\code{specificity}}{the specificity (proportion)}
\item{\code{prev}}{prevalence proportion of the test condition}
\item{\code{ntpp}}{anticipated total tests required per positive result}
\item{\code{prev_untested}}{anticipated prevalence proportion among the untested}
}
}
\description{
\code{ntpp.data.frame} computes the ratio of the total number of
tests performed per positive test result and the anticipated fraction of the
untested (those screened out of testing) who would actually test positive.
}
\details{
The anticipated number of tests required to detect a single positive
\emph{nntp} is given by
\deqn{nntp = (SeP + (1 - Sp)(1 - P)) / SeP}
where \emph{Se} is sensitivity, \emph{P} is prevalence and \emph{Sp} is
specificity. The anticipated prevalence among those screened out is given by
\deqn{Puntested = ((1 - Se)P) / ((1 - Se)P + Sp (1 - P))}
}
