% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntpp-methods.R
\name{ntpp.easy_tool}
\alias{ntpp.easy_tool}
\title{Compute the Ratio of Total Tests to Positive Results from \code{easy_tool} Objects}
\usage{
\method{ntpp}{easy_tool}(object, ..., prev = NULL)
}
\arguments{
\item{object}{an \code{easy_tool}-class object produced by \code{easy_tool}.}

\item{...}{optional arguments to \code{ntpp} methods.}

\item{prev}{an optional prevalence proportion for the test outcome; if missing
the prevalence is obtained from \code{object}.}
}
\value{
A data frame containing the following columns:
\describe{
\item{\verb{sensitivity}}{The sensitivity (proportion) of the screener.}
\item{\verb{specificity}}{The specificity (proportion) of the screener.}
\item{\verb{ntpp}}{the number of tests required to discover
a single positive test result.}
\item{\verb{prev_untested}}{The prevalence proportion of the test
condition among those who are screened out of testing.}
}
}
\description{
\code{ntpp.easy_tool} computes the ratio of the total number of
tests performed per positive test result from \code{easy_tool}-class objects.
}
\details{
The anticipated number of tests required to detect a single positive
\emph{nntp} is given by
\deqn{nntp = (SeP + (1 - Sp)(1 - P)) / SeP}
where \emph{Se} is sensitivity, \emph{P} is prevalence and \emph{Sp} is
specificity. The anticipated prevalence among those screened out is given by
\deqn{Puntested = ((1 - Se)P) / ((1 - Se)P + Sp (1 - P))}
}
\examples{
attach(uniobj1)
tool <- easy_tool(uniobj1, max = 3, crossval = TRUE)
ntpp(tool)

}
