% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticScreening.R
\name{ntpp.logisticScreenr}
\alias{ntpp.logisticScreenr}
\title{\code{ntpp.logisticScreenr} is a method for computation of the anticipated
number of tests per positive test result}
\usage{
\method{ntpp}{logisticScreenr}(object, type = "cvResults", prev = NULL)
}
\arguments{
\item{object}{a \code{logisticScreenr}-class object produced by \code{logisticScreenr}.}

\item{type}{(character) one of \verb{"cvResults"} (the default) or
\verb{"isResults"} to specify \emph{k}-fold cross-validated or in-sample
receiver-operating characteristics, respectively.}

\item{prev}{an optional prevalence proportion for the test outcome; if missing
the prevalence is obtained from \code{object}.}
}
\value{
A data frame containing the following columns:
\describe{
\item{\verb{sensitivity}}{The sensitivity (proportion) of the screener.}
\item{\verb{specificity}}{The specificity (proportion) of the screener.}
\item{\verb{ntpp}}{the number of tests required to discover
a single positive test result.}
\item{\verb{prev_untested}}{The prevalence proportion of the test
condition among those who are screened out of testing.}
}
}
\description{
\code{ntpp.logisticScreenr} is a method for computation of the anticipated
number of tests per positive test result
}
\details{
The anticipated number of tests needed to observe a single positive test
result is a function of sensitivity, specificity and the prevalence proportion
of the condition being tested.
}
\examples{
load(uniobj2)
class(uniobj2)
ntpp(uniobj2)

}
