% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{ntpp}
\alias{ntpp}
\title{\code{ntpp} is an S3 generic function that computes the anticipated number of
tests per positive test result and prevalence among the subjects who would be
screened out of testing)}
\usage{
\code{getWhat(object, ...)}
}
\description{
\code{ntpp} is an S3 generic function that computes the anticipated number of
tests per positive test result and prevalence among the subjects who would be
screened out of testing)
}
\seealso{
\code{link[screenr]{ntpp.glmpathScreenr}}

\code{link[screenr]{ntpp.logisticScreenr}}

\code{link[screenr]{ntpp.data.frame}}
}
