% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntpp-methods.R
\name{ntpp.simple_screenr}
\alias{ntpp.simple_screenr}
\title{Compute the Ratio of Total Tests to Positive Results from \code{simple_screenr} Objects}
\usage{
\method{ntpp}{simple_screenr}(object, ..., prev = NULL)
}
\arguments{
\item{object}{a \code{simple_screenr}-class object produced by \code{simple_screenr}.}

\item{...}{optional arguments to \code{ntpp} methods.}

\item{prev}{an optional prevalence proportion for the test outcome; if missing
the prevalence is obtained from \code{object}.}
}
\value{
A data frame containing the following columns:
\describe{
\item{\verb{sensitivity}}{The sensitivity (proportion) of the screener.}
\item{\verb{specificity}}{The specificity (proportion) of the screener.}
\item{\verb{ntpp}}{the number of tests required to discover
a single positive test result.}
\item{\verb{prev_untested}}{The prevalence proportion of the test
condition among those who are screened out of testing.}
}
}
\description{
\code{ntpp.simple_screenr} computes the ratio of the total number of
tests performed per positive test result from \code{simple_screenr}-class objects.
}
\details{
The anticipated number of tests required to detect a single positive
\emph{nntp} is given by
\deqn{nntp = (SeP + (1 - Sp)(1 - P)) / SeP}
where \emph{Se} is sensitivity, \emph{P} is prevalence and \emph{Sp} is
specificity. The anticipated prevalence among those screened out is given by
\deqn{Puntested = ((1 - Se)P) / ((1 - Se)P + Sp (1 - P))}
}
