% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScreening.R
\name{ntpp.simpleScreenr}
\alias{ntpp.simpleScreenr}
\title{\code{ntpp.simpleScreenr} is a method for computation of the anticipated
number of tests per positive test result}
\usage{
\method{ntpp}{simpleScreenr}(object, prev = NULL)
}
\arguments{
\item{object}{a \code{simpleScreenr}-class object produced by \code{simpleScreenr}.}

\item{prev}{an optional prevalence proportion for the test outcome; if missing
the prevalence is obtained from \code{object}.}
}
\value{
A data frame containing the following columns:
\describe{
\item{\verb{sensitivity}}{The sensitivity (proportion) of the screener.}
\item{\verb{specificity}}{The specificity (proportion) of the screener.}
\item{\verb{ntpp}}{the number of tests required to discover
a single positive test result.}
\item{\verb{prev_untested}}{The prevalence proportion of the test
condition among those who are screened out of testing.}
}
}
\description{
\code{ntpp.simpleScreenr} is a method for computation of the anticipated
number of tests per positive test result
}
\details{
The anticipated number of tests needed to observe a single positive test
result is a function of sensitivity, specificity and the prevalence proportion
of the condition being tested.
}
\examples{
data(unicorns)
toosimp <- simpleScreenr(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5 + Q6, data = unicorns)
ntpp(simple)

}
