% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialScreening.R
\name{plot.binomscreenr}
\alias{plot.binomscreenr}
\title{An S3 Plot Method for \code{binomscreenr} Objects}
\usage{
\method{plot}{binomscreenr}(obj,
  main = "Receiver Operating Characteristics", ...)
}
\arguments{
\item{obj}{An object of class \code{binomscreenr} produced by function
\code{binomialScreening}}
}
\value{
A \code{lattice} graphical object
}
\description{
This is essentially a wrapper function for \code{lattice::xyplot} customized
to produce a graph of Receiver Operating Characteristic curves.
}
