% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.easy_tool}
\alias{plot.easy_tool}
\title{Plot ROC Curves from \code{easy_tool}-Class Objects}
\usage{
\method{plot}{easy_tool}(
  x,
  ...,
  plot_ci = TRUE,
  conf_level = 0.95,
  bootreps = 4000,
  print_auc = TRUE,
  partial_auc = c(0.8, 1),
  partial_auc_focus = c("sensitivity", "specificity"),
  partial_auc_correct = TRUE,
  type = "S"
)
}
\arguments{
\item{x}{an object of class \code{easy_tool}.}

\item{...}{any additional arguments passed to \code{pROC::plot.roc} or
\code{pROC::lines.roc}.}

\item{plot_ci}{(logical) plot confidence intervals if \verb{TRUE}.}

\item{conf_level}{confidence level}

\item{bootreps}{the number of bootstrap replications for estimation of
confidence intervals. Default: 4000.}

\item{print_auc}{logical indicator for printing the area
under the ROC curve (AUC) on the plot.  Default: \verb{TRUE}.}

\item{partial_auc}{One of \verb{FALSE} or a length two numeric vector
of the form \verb{c(a, b)} where \verb{a} and \verb{b} are the endpoints
of the interval over which to compute the partial AUC (pAUC). Ignored if
\code{print_auc = FALSE}. Default: \code{c(0.8, 1)}.}

\item{partial_auc_focus}{one of \verb{"sensitivity"} or \verb{"specificity"},
indicating the measure for which the partial AUC is to be computed. Default:
\verb{"specificity"}.}

\item{partial_auc_correct}{logical indictor for transformation of the pAUC
to fall within the range from 0.5 (random guess) to 1.0 (perfect
classification). Default: \verb{TRUE}.}

\item{type}{type of plot. See \code{\link[base]{plot}}. Default: \verb{"S"}.}
}
\value{
This function produces a plot as a side effect and (optionally)
returns a dataframe containing sensitivities, specificities and their
lower and upper confidence limits for threshold values of Pr(response = 1).
}
\description{
\code{plot.easy_tool} plots the \emph{k}-fold cross-validated
receiver-operating characteristics, including confidence intervals on the
combinations of the local maxima of sensitivity and specificity.
}
\details{
\code{plot.easy_tool} is an enhanced convenience wrapper for
\code{pROC::plot.roc}.
}
\examples{
attach(uniobj1)
tool <- easy_tool(uniobj1, max = 3, crossval = TRUE)
plot(tool)
}
\references{
Fawcett T. An introduction to ROC analysis. Pattern Recognition Letters. 2006.
27(8):861-874. \url{https://doi.org/10.1016/j.patrec.2005.10.010}

Linden A. Measuring diagnostic and predictive accuracy in disease
management: an introduction to receiver operating characteristic (ROC) analysis.
Journal of Evaluation in Clinical Practice. 2006; 12(2):132-139.
\url{https://onlinelibrary.wiley.com/doi/epdf/10.1111/j.1365-2753.2005.00598.x}

Robin X, Turck N, Hainard A, Tiberti N, Lisacek F, Sanchez J-C, Muller M.
pROC: an open-source package for R and S+ to analyze and compare ROC curves.
BMC Bioinformatics 2011; 12:77. \url{https://www.biomedcentral.com/1471-2105/12/77}
}
