% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{plot.ROC}
\alias{plot.ROC}
\title{An S3 Plot Method for \code{ROC} Objects}
\usage{
\method{plot}{ROC}(x)
}
\arguments{
\item{x}{An object of class ROC as produced by \code{simpleScreening()}}
}
\description{
An S3 Plot Method for \code{ROC} Objects
}
