% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_lost.R
\name{plot_barcode_lost_for_gene}
\alias{plot_barcode_lost_for_gene}
\title{Plot number of barcode lost for gene}
\usage{
plot_barcode_lost_for_gene(screenR_Object)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}
}
\value{
return a tibble containing the number of mapped read for sample
}
\description{
This function plots the number of barcode lost in each sample
}
