% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode_trend.R
\name{plot_barcode_trend}
\alias{plot_barcode_trend}
\title{Plot the trend over time of the Barcode}
\usage{
plot_barcode_trend(
  list_data_measure,
  genes,
  n_col = 2,
  size_line = 1,
  color = NULL
)
}
\arguments{
\item{list_data_measure}{A list containing the measure table of the
different time point}

\item{genes}{vector of genes name}

\item{n_col}{The number of col to use in the facette wrap}

\item{size_line}{the dimension of the line}

\item{color}{vector of colors}
}
\value{
The the plot of the
}
\description{
Plot the log2FC over time of the barcode in the different
time point
}
