% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{plotROC}
\alias{plotROC}
\title{Plot Receiver Operating Characteristics}
\usage{
plotROC(x, main = "Receiver Operating Characteristic", ...)
}
\arguments{
\item{x}{an object of class \code{binomscreener} \code{simplescreenr}.}

\item{main}{plot title.}

\item{...}{arguments to be passed to or from \code{lattice::xyplot}.}
}
\value{
A \code{lattice} graphical object.
}
\description{
Plot Receiver Operating Characteristics
}
