% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmpathScreening.R
\name{predict.glmpathScreenr}
\alias{predict.glmpathScreenr}
\title{\code{predict.glmpathScreenr} is a prediction method for objects of class \code{glmpathScreenr}}
\usage{
\method{predict}{glmpathScreenr}(object = NULL, newdata = NULL)
}
\arguments{
\item{object}{an object of class \code{glmpathScreenr} produced by
\code{\link[screenr]{glmpathScreenr}}.}

\item{newdata}{new dataframe from which predicted probabilities of positive test results are desired.
The dataframe must contain values of the same response variables and covariates that were used
to obtain \code{obj}.}
}
\value{
\code{predict.glmpathScreenr} returns (invisibly) a dataframe augmenting the complete cases
in \code{newdata} with the predicted probabilities of positive test results \code{phat_minAIC} and
\code{phat_minBIC} from the models that produced the minimum AIC and BIC, respectively.
}
\description{
\code{predict.glmpathScreenr} is a prediction method for objects of class \code{glmpathScreenr}
}
\details{
This method is a convenience wrapper for \code{link[glmpath]{predict.glmpath}}.
}
\examples{
attach(uniobj1)
## Get some new observations
new_corns <- data.frame(ID = c("Alice D.", "Bernie P."),
                        testresult = c(NA, NA),
                        Q1 = c(0, 0), Q2 = c(0, 0), Q3 = c(0, 1), Q4 = c(0, 0),
                        Q5 = c(0, 1), Q6 = c(1, 1))
## Predict the probabilities of testing positive for the new subjects
predict(uniobj, new_corns)
}
