% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-methods.R
\name{predict.lasso_screenr}
\alias{predict.lasso_screenr}
\title{A Prediction Method for \code{lasso_screenr}-Class Objects}
\usage{
\method{predict}{lasso_screenr}(object = NULL, ..., newdata = NULL)
}
\arguments{
\item{object}{an object of class \code{lasso_screenr} produced by
\code{`lasso_screenr`}.}

\item{...}{optional arguments to \code{predict} methods.}

\item{newdata}{new dataframe from which predicted probabilities of positive
test results are desired.  The dataframe must contain values of the same
response variables and covariates that were used to obtain \code{obj}.}
}
\value{
\code{predict.lasso_screenr} returns (invisibly) a dataframe
augmenting the complete cases in \code{newdata} with the predicted
probabilities of positive test results \code{phat_minAIC} and
\code{phat_minBIC} from the models that produced the minimum AIC and BIC,
respectively.
}
\description{
\code{predict.lasso_screenr} computes predicted probabilities of
positive test results from new data.
}
\details{
This method is a convenience wrapper for \
code{`glmpath::predict.glmpath`}.
}
\examples{
attach(uniobj1)
## Get some new observations
new_corns <- data.frame(ID = c("Alice D.", "Bernie P."),
                        testresult = c(NA, NA),
                        Q1 = c(0, 0), Q2 = c(0, 0), Q3 = c(0, 1), Q4 = c(0, 0),
                        Q5 = c(0, 1), Q6 = c(0, 1), Q7 = c(0, 1))
## Predict the probabilities of testing positive for the new subjects
print(predict(uniobj1, newdata = new_corns))
}
