% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticScreening.R
\name{predict.logisticScreenr}
\alias{predict.logisticScreenr}
\title{\code{predict.logisticScreenr} is an S3 prediction method for objects of class \code{logisticScreenr}}
\usage{
\method{predict}{logisticScreenr}(object = NULL, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{logisticScreenr} produced by
\code{\link[screenr]{logisticScreenr}}.}

\item{newdata}{new dataframe from which predicted probabilities of positive test results are
desired. The dataframe must contain values of the same response variables and covariates that
were used to obtain \code{object}.}
}
\value{
\code{predict.logisticScreenr} returns (invisibly) a dataframe augmenting
\code{newdata} with the predicted probabilities of positive test results \code{phat}.
}
\description{
\code{predict.logisticScreenr} is an S3 prediction method for objects of class \code{logisticScreenr}
}
\details{
This method is a convenience wrapper for \code{link[stats]{predict.glm}}.
}
\examples{
load(uniobj2)
class(uniobj2)
## Get some new observations
new_corns <- data.frame(ID = c("Alice D.", "Bernie P."),
                        testresult = c(NA, NA), Q1 = c(0, 0), Q2 = c(0, 0),
                        Q3 = c(0, 0), Q4 = c(0, 0), Q5 = c(0, 1), Q6 = c(0, 1 ))
## Predict the probabilities of testing positive for the new subjects
predict(uniobj2, newdata = new_corns)
}
