% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-methods.R
\name{predict.logreg_screenr}
\alias{predict.logreg_screenr}
\title{A Prediction Method for \code{logreg_screenr}-Class Objects}
\usage{
\method{predict}{logreg_screenr}(object = NULL, ..., newdata = NULL)
}
\arguments{
\item{object}{an object of class \code{logreg_screenr} produced by
\code{`logreg_screenr`}.}

\item{...}{optional arguments to \code{predict} methods.}

\item{newdata}{new dataframe from which predicted probabilities of positive
test results are desired. The dataframe must contain values of the same
response variables and covariates that were used to obtain \code{object}.}
}
\value{
\code{predict.logreg_screenr} returns (invisibly) a dataframe
augmenting \code{newdata} with the predicted probabilities of positive test
results \code{phat}.
}
\description{
\code{predict.logreg_screenr} computes predicted probabilities
of positive test results from new data.
}
\details{
This method is a convenience wrapper for \code{`stats::predict.glm`}.
}
\examples{
attach(uniobj2)
## Get some new observations
new_corns <- data.frame(ID = c("Alice D.", "Bernie P."),
                        testresult = c(NA, NA), Q1 = c(0, 0), Q2 = c(0, 0),
                        Q3 = c(0, 0), Q5 = c(0, 1), Q6 = c(0, 1 ),
                        Q7 = c(0, 1))
## Predict the probabilities of testing positive for the new subjects
print(predict(uniobj2, newdata = new_corns))
}
