% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialScreening.R
\name{print.binomscreenr}
\alias{print.binomscreenr}
\title{An S3 Print Method for \code{binomscreenr} Objects.}
\usage{
\method{print}{binomscreenr}(obj)
}
\arguments{
\item{obj}{An object of class \code{binomscreenr} produced by function
\code{binomialScreening}}
}
\description{
An S3 Print Method for \code{binomscreenr} Objects.
}
