% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.easy_tool}
\alias{print.easy_tool}
\title{A Print Method for \code{easy_tool}-Class Objects}
\usage{
\method{print}{easy_tool}(x, ...)
}
\arguments{
\item{x}{an object of class \code{easy_tool}.}

\item{...}{optional arguments to \code{print} methods.}
}
\description{
\code{print.easy_tool} is a print method.
}
\examples{
attach(uniobj1)
print(uniobj1)
}
\seealso{
\code{get_what.easy_tool(from, what)} for \code{what =} \verb{"ROCci"}.
}
