% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyTool.R
\name{print.easyTool}
\alias{print.easyTool}
\title{\code{print.easyTool} is a print method for \code{easyTool} objects}
\usage{
\method{print}{easyTool}(object)
}
\arguments{
\item{object}{an object of class \code{easyTool}}
}
\description{
\code{print.easyTool} is a print method for \code{easyTool} objects
}
\examples{
load(uniobj1)
class(uniobj1)
print(uniobj1)
}
