% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.glmpathScreenr}
\alias{print.glmpathScreenr}
\title{code{print.glmpathScreenr} is a print method for \code{glmpathScreenr} objects}
\usage{
\method{print}{glmpathScreenr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{glmpathScreenr}}

\item{...}{optional arguments to \code{print} methods.}
}
\description{
code{print.glmpathScreenr} is a print method for \code{glmpathScreenr} objects
}
\examples{
attach(uniobj1)
print(uniobj1)
}
