% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.lasso_screenr}
\alias{print.lasso_screenr}
\title{An S3 Print Method for \code{screenr} Objects}
\usage{
\method{print}{lasso_screenr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{lasso_screenr}}

\item{...}{optional arguments to \code{print} methods.}
}
\description{
\code{print.lasso_screenr} is a print method for
\code{lasso_screenr}-class objects.
}
\examples{
attach(uniobj1)
print(uniobj1)
}
