% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.simple_screenr}
\alias{print.simple_screenr}
\title{\code{print.simple_screenr} is an S3 print method for \code{simple_screenr} objects.}
\usage{
\method{print}{simple_screenr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{simplescreenr}.}

\item{...}{optional arguments to \code{print} methods.}
}
\value{
Nothing. Thresholds, specificities and sensitivities are printed as a
side effect.
}
\description{
\code{print.simple_screenr} is an S3 print method for \code{simple_screenr} objects.
}
\examples{
data(unicorns)
toosimple <- simple_screenr(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5 + Q6,
                          data = unicorns)
print(toosimple)

}
