% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.simpleScreenr}
\alias{print.simpleScreenr}
\title{\code{print.simpleScreenr} is an S3 print method for \code{simpleScreenr} objects.}
\usage{
\method{print}{simpleScreenr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{simplescreenr}.}

\item{...}{optional arguments to \code{print} methods.}
}
\value{
Nothing. Thresholds, specificities and sensitivities are printed as a
side effect.
}
\description{
\code{print.simpleScreenr} is an S3 print method for \code{simpleScreenr} objects.
}
\examples{
data(unicorns)
toosimple <- simpleScreenr(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5 + Q6,
                          data = unicorns)
print(toosimple)

}
