% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{rescale_to_Int}
\alias{rescale_to_Int}
\title{Rescale a strictly positive vector of real numbers to integers}
\usage{
rescale_to_Int(x, max)
}
\arguments{
\item{x}{numeric vector of positive real numbers.}

\item{max}{the value of largest element in the rescaled integer-valued
vector.}
}
\value{
A vector of integers corresponding to \code{x} in which smallest
element is 1 and the largest element is \code{max}
}
\description{
A convenience wrapper function which rescales a strictly positive (all
elements are greater than zero) vector and rescales it to a vector of
integers.
}
\seealso{
\code{\link[scales]{rescale}}
}
