% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_lm.R
\name{screen_lm}
\alias{screen_lm}
\title{screen_lm}
\usage{
screen_lm(df, responses, factors)
}
\arguments{
\item{df, }{data.frame, a \code{data.frame} with factors and responses as columns}

\item{responses, }{character, a vector of characters corresponding to column names
in \code{df} to be used as responses}

\item{factors, }{character, a vector of characters corresponding to column names
in \code{df} to be used as factors}
}
\value{
A list with the results data frame in \code{results} and inputs in \code{input}.
}
\description{
Use lm to screen numeric variables to numeric factors
}
\details{
\code{screen_lm} uses \code{lm} to screen responses against factors one to one
}
\examples{
out <- screen_lm(mtcars, c("cyl", "disp"), c("wt"))

}
