% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenr-class.R
\docType{class}
\name{screenr_object-class}
\alias{screenr_object-class}
\alias{screenr_object}
\title{S4 ScreenR object Class
The screenr_object class is the main object of the package, it is passed
to a series of function to perform the analysis. I}
\description{
S4 ScreenR object Class
The screenr_object class is the main object of the package, it is passed
to a series of function to perform the analysis. I
}
\section{Slots}{

\describe{
\item{\code{count_table}}{It is used to store the count table to perform the
analysis}

\item{\code{annotation_table}}{It is used to store the annotation of the shRNA}

\item{\code{groups}}{It is used to store the vector of treated and untreated}

\item{\code{replicates}}{It is used to store information about the replicates}

\item{\code{normalized_count_table}}{It is used to store a normalized verision of
the count table}

\item{\code{data_table}}{It is used to store a tidy format of the count table}
}}

\examples{
data("count_table", package = "ScreenR")
data("annotation_table", package = "ScreenR")

groups <- factor(c("T1/T2", "T1/T2", "Treated", "Treated", "Treated",
                   "Control", "Control", "Control", "Treated", "Treated",
                   "Treated", "Control", "Control", "Control"))

obj <- create_screenr_object(table = count_table,
                             annotation = annotation_table,
                             groups = groups,
                             replicates = c(""))
}
\concept{objects}
