% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{se_sp_max}
\alias{se_sp_max}
\title{Return dataframe rows for which specificity is the maximum for each sensitivity}
\usage{
se_sp_max(object)
}
\arguments{
\item{object}{a dataframe containing at least columns named \code{sensitivity}
and \code{specificity}}
}
\value{
a dataframe which is a subset of \code{object} containing only those
rows for which specificity was the maximum for each unique value of sensitivity.
}
\description{
Given a dataframe containing multiple values of specificity for
each value of sensitivity, return only the rows containing the largest
specificity for each unique value of sensitivity.
}
