% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera_method.R
\name{select_number_barcode}
\alias{select_number_barcode}
\title{Select  number of Barcode}
\usage{
select_number_barcode(gene, genesymbols, number_barcode)
}
\arguments{
\item{gene}{The gene name}

\item{genesymbols}{The gene symbols  list}

\item{number_barcode}{The number of barcode to select}
}
\value{
The barcode of the gene passed as input
}
\description{
Compute a unique gene symbol for gene
}
\keyword{internal}
