% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sens_spec_plus}
\alias{sens_spec_plus}
\title{Compute Sensitivity, specificity and a few friends}
\usage{
sens_spec_plus(
  test = NULL,
  gold = NULL,
  data = NULL,
  method = c("exact", "wilson", "agresti", "clopper-pearson", "jeffreys"),
  conf_level = 0.95
)
}
\arguments{
\item{test}{character-valued name of the variable containing testing results,
coded as 0 for negative and 1 for positive.}

\item{gold}{character-valued name of the variable containing gold standard,
coded as 0 for negative and 1 for positive.}

\item{data}{data frame containing \code{test} and \code{gold}.}

\item{method}{type of confidence interval
(\verb{"exact", "wilson", "agresti", "clopper-pearson" or"jeffreys"}). Default:
\verb{"exact"}.}

\item{conf_level}{confidence level, a numeric value between 0 and 1.
Default: 0.95.}
}
\value{
a list containing components \code{table} and \code{ests}:
\describe{
\item{\code{table}}{a 2 x 2 table which is the anti-transpose of the
result produced by \code{base::table(gold, test)}.}
\item{\code{ests}}{a dataframe containing the apparent and true positive
proportions, sensitivity, specificity, positive predictive value (PPV),
negative predictive value (NPV), and the lower and upper confidence limits for
each.}
}
}
\description{
\code{sens_spec_plus} computes sensitivity, specificity and a
few friends from a gold standard and testing results. \code{sens_spec_plus}
is a convenience wrapper for \code{epiR::epi.tests}.
}
\examples{
Gold <- rbinom(20, 1, 0.50)
Test <- Gold; Test[c(3, 5, 9, 12, 16)] <- 1 - Test[c(3, 5, 9, 12, 16)]
dat <- data.frame(Gold = Gold, Test = Test)
sens_spec_plus(test = "Test", gold = "Gold", data = dat)
}
\seealso{
\code{\link[epiR]{epi.tests}}
}
