% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenr-class.R
\name{set_annotation_table,screenr_object-method}
\alias{set_annotation_table,screenr_object-method}
\alias{set_annotation_table,screenr_object}
\title{Set ScreenR annotation table}
\usage{
\S4method{set_annotation_table}{screenr_object}(object, annotation_table)
}
\arguments{
\item{object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}

\item{annotation_table}{a table containing the annotation for each shRNA}
}
\value{
The ScreenR object with the annotation table
}
\description{
Set function for the annotation table of the ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
annotation <- get_annotation_table(object)
set_annotation_table(object, annotation)
}
\concept{objects}
