% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenr-class.R
\name{set_count_table}
\alias{set_count_table}
\title{Set ScreenR count table}
\usage{
set_count_table(object, count_table)
}
\arguments{
\item{object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}

\item{count_table}{A count table containing in each row an shRNA and in each
column a sample}
}
\value{
The ScreenR object with the count table
}
\description{
Set function for the count table of the ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
counts <- get_count_table(object)
set_count_table(object, counts)
}
\concept{objects}
