% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/screenr-class.R
\name{set_data_table}
\alias{set_data_table}
\alias{set_data_table,screenr_object-method}
\alias{set_data_table,screenr_object}
\title{Set ScreenR data_table}
\usage{
set_data_table(object, data_table)

\S4method{set_data_table}{screenr_object}(object, data_table)
}
\arguments{
\item{object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}

\item{data_table}{A count table in a tidy format}
}
\value{
The ScreenR object with the setted data_table
}
\description{
Set function for the data_table of the ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
data_table <- get_data_table(object)
set_data_table(object, data_table)
}
\concept{objects}
