% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenr-class.R
\name{set_groups}
\alias{set_groups}
\title{Set ScreenR groups}
\usage{
set_groups(object, groups)
}
\arguments{
\item{object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}

\item{groups}{The treatment and control groups}
}
\value{
The ScreenR object containing the group field
}
\description{
Set function for the groups of the ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
groups <- get_groups(object)
set_groups(object, groups)
}
\concept{objects}
