% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenr-class.R
\name{set_normalized_count_table}
\alias{set_normalized_count_table}
\title{Set ScreenR normalized_count_table}
\usage{
set_normalized_count_table(object, normalized_count_table)
}
\arguments{
\item{object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}

\item{normalized_count_table}{A table of the normalized count table}
}
\value{
The ScreenR object with the setted normalized_count_table
}
\description{
Set function for the normalized_count_table of
             the ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
normalized_count_table <- get_normalized_count_table(object)
normalized_count_table
set_normalized_count_table(object, normalized_count_table)
}
\concept{objects}
