% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenr-class.R
\name{set_replicates}
\alias{set_replicates}
\title{Set ScreenR replicates}
\usage{
set_replicates(object, replicates)
}
\arguments{
\item{object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}

\item{replicates}{The vecotr containing the replicates name}
}
\value{
The ScreenR object with the specific replicates
}
\description{
Set function for the replicates of the ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
replicates <- get_replicates(object)
set_replicates(object, replicates)
}
\concept{objects}
