% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_screenr.R
\name{simple_screenr}
\alias{simple_screenr}
\title{An Overly Simple Approach to Test Screening}
\usage{
simple_screenr(
  formula,
  data,
  partial_auc = c(0.8, 1),
  partial_auc_focus = "sensitivity",
  partial_auc_correct = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}} defining the
testing outcome and predictor covariates.}

\item{data}{the "training" sample; a data frame containing the testing outcome
and predictive covariates to be used for testing screening.  The testing
outcome must be binary (0,1) indicating negative and positive test results,
respectively, or logical (\verb{TRUE}/\verb{FALSE}), and the screening scores
are the row-wise sums of the values of those covariates.  The covariates are
typically binary (0 = no, 1 = yes) responses to questions, but the responses
may also be ordinal numeric values.}

\item{partial_auc}{either a logical \verb{FALSE} or a numeric vector of the
form \code{c(left, right)} where left and right are numbers in the interval
[0, 1] specifying the endpoints for computation of the partial area under the
ROC curve (pAUC). The total AUC is computed if \code{partial\_auc} = \verb{FALSE}.
Default: \code{c(0.8, 1.0)}}

\item{partial_auc_focus}{one of \verb{"sensitivity"} or \verb{specificity},
specifying for which the pAUC should be computed.  \code{partial.auc.focus} is
ignored if \code{partial\_auc} = \verb{FALSE}.  Default: \verb{"sensitivity"}.}

\item{partial_auc_correct}{logical value indicating whether the pAUC should be
transformed the interval from 0.5 to 1.0. \code{partial\_auc\_correct} is
ignored if \code{partial\_auc} = \verb{FALSE}. Default: \verb{TRUE}).}

\item{conf_level}{a number between 0 and 1 specifying the confidence level
for confidence intervals for the (partial)AUC. Default: 0.95.}
}
\value{
An object of class \code{simple_screenr} containing the elements:
\describe{
\item{\code{Call}}{The function call.}
\item{\code{Prevalence}}{Prevalence of the test condition in the training sample.}
\item{\code{ISroc}}{An object of class \code{\link[pROC]{roc}} containing
the "in-sample" (overly-optimistic) receiver operating characteristics,
and additional functions for use with this object are available in the
\code{pROC} package.}
\item{\code{Scores}}{The training sample, including the scores.}
}
}
\description{
\code{simple_screenr} implements the method described in
Bandason et al. (2016).
}
\details{
\code{simple_screenr} computes the in-sample (\emph{overly optimistic})
performances for development of a very simple test screening tool based on
the sums of affirmative questionnaire responses.  \code{simpleScreener} is
not optimized and is intended only for comparision with \code{lasso_screenr} or
\code{logreg_screenr}, either of which will almost certainly out-perform
\code{simple_screenr}.
}
\examples{
data(unicorns)
toosimple <- simple_screenr(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5 + Q6 + Q7,
                           data = unicorns)
summary(toosimple)

}
\references{
Bandason T, McHugh G, Dauya E, Mungofa S, Munyati SM, Weiss HA, Mujuru H,
Kranzer K, Ferrand RA. Validation of a screening tool to
identify older children living with HIV in primary care facilities in high
HIV prevalence settings. AIDS. 2016;30(5):779-785
\url{http://dx.doi.org/10.1097/QAD.0000000000000959}

Robin X, Turck N, Hainard A, Tiberti N, Lisacek F, Sanchez J-C,
M昼㹣ller M. \code{pROC}: An open-source package for \code{R} and S+ to
analyze and compare ROC curves. BMC Bioinformatics. 2011;12(77):1-8.
\url{http://doi.org/10.1186/1471-2105-12-77}
}
\seealso{
\code{\link{easy_tool}} for a better approach to simplification
using the results from \code{lasso_screenr} or \code{logreg_screenr}.

\code{\link{lasso_screenr}}, \code{\link{logreg_screenr}}
}
