% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.easyTool}
\alias{summary.easyTool}
\title{\code{summary.easyTool} returns a summary of the receiver-operating
characteristic produced by \code{easyTool}}
\usage{
\method{summary}{easyTool}(object, ...)
}
\arguments{
\item{object}{an \code{easyTool} object.}

\item{...}{optional arguments passed to \code{summary} methods.}
}
\value{
a dataframe containing the summary, including the Df, Deviance,
AIC and BIC for each step along the GLM path for which the active set
changed.
}
\description{
\code{summary.easyTool} returns a summary of the receiver-operating
characteristic produced by \code{easyTool}
}
\details{
This is essentially a wrapper for \code{glmpath::summary.glmpath}
provided for \code{glmpathScreenr} objects.
}
\examples{
attach(uniobj1)
summary(uniobj1)
}
