% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\name{summary.gee_screenr}
\alias{summary.gee_screenr}
\title{An S3 Summary Method for \code{screenr} Objects}
\usage{
\method{summary}{gee_screenr}(object, ..., diagnostics = FALSE)
}
\arguments{
\item{object}{an object of class \code{gee_screenr} produced by function
\code{gee_screenr}.}

\item{...}{optional arguments passed to \code{summary} methods.}

\item{diagnostics}{a logical value; plot model diagnostics if \verb{TRUE}.}
}
\value{
Nothing.  Summaries are printed as a side effect.
}
\description{
\code{summary.gee_screenr} provides a summary method for
\code{gee_screenr}-class objects.
}
\examples{
attach(uniobj2)
summary(uniobj2)
}
