% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmpathScreening.R
\name{summary.glmpathScreenr}
\alias{summary.glmpathScreenr}
\title{\code{summary.glmpathScreenr} returns a summary of the GLM path regularizer}
\usage{
\method{summary}{glmpathScreenr}(object)
}
\arguments{
\item{object}{a glmpathScreenr object}
}
\value{
a dataframe containing the summary, including the Df, Deviance,
AIC and BIC for each step along the GLM path for which the active set
changed.
}
\description{
\code{summary.glmpathScreenr} returns a summary of the GLM path regularizer
}
\details{
This is essentially a wrapper for \code{glmpath::summary.glmpath}
provided for \code{glmpathScreenr} objects.
}
\examples{
attach(uniobj1)
summary(uniobj1)
}
