% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\name{summary.lasso_screenr}
\alias{summary.lasso_screenr}
\title{A Summary Method for \code{lasso_screenr}-Class Objects}
\usage{
\method{summary}{lasso_screenr}(object, ...)
}
\arguments{
\item{object}{a lasso_screenr object}

\item{...}{optional arguments passed to \code{summary} methods.}
}
\value{
a dataframe containing the summary, including the Df, Deviance,
AIC and BIC for each step along the GLM path for which the active set
changed.
}
\description{
\code{summary.lasso_screenr} provides a summary method for
\code{lasso_screenr}-class objects.
}
\details{
This is essentially a wrapper for \code{glmpath::summary.glmpath}
provided for \code{lasso_screenr} objects.
}
\examples{
attach(uniobj1)
summary(uniobj1)
}
