% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.logisticScreenr}
\alias{summary.logisticScreenr}
\title{\code{summary.logisticScreenr} is an S3 print method for \code{logisticScreenr}
objects.}
\usage{
\method{summary}{logisticScreenr}(object, ..., diagnostics = FALSE)
}
\arguments{
\item{object}{an object of class \code{logisticScreenr} produced by function
\code{logisticScreenr}.}

\item{...}{optional arguments passed to \code{summary} methods.}

\item{diagnostics}{a logical value; plot model diagnostics if \verb{TRUE}.}
}
\value{
Nothing.  Summaries are printed as a side effect.
}
\description{
\code{summary.logisticScreenr} is an S3 print method for \code{logisticScreenr}
objects.
}
\examples{
attach(uniobj2)
summary(uniobj2)
}
