% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\name{summary.simple_screenr}
\alias{summary.simple_screenr}
\title{A Summary Method for \code{simple_screenr}-Class Objects}
\usage{
\method{summary}{simple_screenr}(object, ...)
}
\arguments{
\item{object}{an object of class \code{simple_screenr}.}

\item{...}{optional arguments passed to \code{summary} methods.}
}
\value{
Nothing.  Thresholds, specificities and sensitivities are printed as
a side effect.
}
\description{
\code{summary.simple_screenr} provides a summary method for
\code{simple_screenr}-class objects.
}
\examples{
data(unicorns)
toosimple <- simple_screenr(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5 + Q6 + Q7,
                            data = unicorns)
summary(toosimple)
}
