% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.simpleScreenr}
\alias{summary.simpleScreenr}
\title{\code{summary.simpleScreenr} is a summary method for \code{simpleScreenr} objects}
\usage{
\method{summary}{simpleScreenr}(object, ...)
}
\arguments{
\item{object}{an object of class \code{simpleScreenr}.}

\item{...}{optional arguments passed to \code{summary} methods.}
}
\value{
Nothing.  Thresholds, specificities and sensitivities are printed as
a side effect.
}
\description{
\code{summary.simpleScreenr} is a summary method for \code{simpleScreenr} objects
}
\examples{
data(unicorns)
toosimple <- simpleScreenr(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5 + Q6,
                          data = unicorns)
summary(toosimple)
}
