% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_Annotation.R
\docType{data}
\name{Table_Annotation}
\alias{Table_Annotation}
\title{Table for the annotation of Barcode}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{Gene}{It Contains the gene name}

  \item{Gene_ID}{It Contains a unique ID of the Gene}

  \item{Barcode}{It contains an ID that identify each barcode, it can be use
  to marge the annotation table with the count table}

  \item{Sequence}{It contains the cDNA sequence of the Barcode}

  \item{Library}{It contains the library from which the Barcode come from}
}
}
\usage{
data(Table_Annotation)
}
\description{
Table for the annotation of Barcode
}
\keyword{datasets}
