% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unicorns}
\alias{unicorns}
\title{UIV Testing Data on Unicorns}
\format{A data frame with six columns:
\describe{
\item{\code{testresult}}{UIV status, where 0 and 1 denote negative and positive test results, repectively.}
\item{\code{Q1--Q5}}{Responses (0 = "no", 1 = "yes") to screening questions 1--5.}
}}
\usage{
unicorns
}
\description{
A preliminary study was conducted in which a random sample of 5,000 properly
consented unicorns were each asked five questions about their behavior and
health. They responded by stomping a hoof once to indicate "no", and twice to
indicate "yes".  A sample of blood was drawn from each, and was subsequently
tested for the presence of antibodies to Unicorn Immunodeficiency Virus (UIV)
using a standard assay algorithm.
}
\examples{
\dontrun{
unicorns
}
}
\keyword{datasets}
