% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniobj2-data.R
\docType{data}
\name{uniobj2}
\alias{uniobj2}
\title{A \code{logreg_screenr object}}
\format{
An object of class \code{logreg_screenr}
}
\usage{
uniobj2
}
\description{
The result of \code{uniobj2 <- logreg_screenr(testresult ~ Q1 + Q2 + Q3 +
Q4 + Q5 + Q6 + Q7, data = unicorns, link = "logit", Nfolds = 10, seed = 123)}
}
\examples{
\dontrun{
summary(uniobj2)
}
}
\keyword{logistic}
\keyword{ordinary}
\keyword{regression}
\keyword{screenr}
