% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera_method.R
\name{unique_gene_symbols}
\alias{unique_gene_symbols}
\title{Unique gene Symbols}
\usage{
unique_gene_symbols(gene_symbols, number_barcode = 3)
}
\arguments{
\item{gene_symbols}{The gene symbols list}

\item{...}{
  Arguments passed on to \code{\link[=find_camera_hit]{find_camera_hit}}
  \describe{
    \item{\code{number_barcode}}{Number of barcode that as to be differentially
expressed (DE)in order to consider the gene associated
DE. Example a gene is associated
with 10 shRNA we consider a gene DE if it has at least
number_barcode = 5 shRNA DE.}
  }}
}
\value{
A list of unique gene symbols
}
\description{
Compute a unique gene symbol for gene
}
\keyword{internal}
