% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzv.R
\name{variance_check}
\alias{variance_check}
\title{variance_check}
\usage{
variance_check(nameVec, data)
}
\arguments{
\item{nameVec}{character vector of variables to check}

\item{data}{the dataset}
}
\value{
gives a character vector of variables with high variance
}
\description{
checks the variance of the given variables and discards them if they have zero or near-zero variance
}
\details{
Uses the \code{\link[https://www.rdocumentation.org/packages/caret/versions/6.0-92/topics/nearZeroVar]{nearZeroVar}} near zero variance function from the caret package
}
