% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{z_score}
\alias{z_score}
\title{Z-score function}
\usage{
z_score(screenR_Object)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}.}
}
\value{
return a tibble containing the number of mapped read for sample
}
\description{
Create a boxplot to show the distribution of the mapped reads
             in different sample
}
