% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barcode_creation_Dataframe.R
\name{Barcode_creation_Dataframe}
\alias{Barcode_creation_Dataframe}
\title{Create the table containing the Mapped Sample f}
\usage{
Barcode_creation_Dataframe(path_file)
}
\arguments{
\item{path_file}{Path of the file}
}
\value{
return the table containing barcode that have mapped to the samples
}
\description{
This function take as input the path of a file creating using
             the samtools idxstats
}
