% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_lost.R
\name{barcode_lost}
\alias{barcode_lost}
\title{Count Barcode Lost}
\usage{
barcode_lost(screenR_Object)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}
}
\value{
return a tibble containing the number of barcode lost for sample
}
\description{
This fuction count the number of Barcode lost during the
             sequenceing. A barcode is lost if it has zero mapped read
}
