% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_lost.R
\name{barcode_lost}
\alias{barcode_lost}
\title{Count Barcode Lost}
\usage{
barcode_lost(screenR_Object)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}
}
\value{
Return a tibble containing the number of barcodes lost for each
        sample
}
\description{
This fuction counts the number of Barcode lost during the
             sequencing. A barcode is lost if it has zero mapped read.
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))

# In order to count the number of barcode lost just the ScreenR object is
# needed
head(barcode_lost(object))

}
\concept{compute}
