% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialScreening.R
\name{binomialScreening}
\alias{binomialScreening}
\title{Test-Screening Tool Based on Binomial Regression}
\usage{
binomialScreening(formula, data = NULL, link = "logit", Nfolds = 40L, ...)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}  defining the testing
outcome and predictor covariates, which is passed to \code{stats::glm()}.}

\item{data}{the "training" sample; a data frame containing the testing
outcome and predictive covariates to be used for testing screening.  The
testing outcome must be binary (0,1) indicating negative and positive test
results, respectively, or logical (\verb{TRUE}/\verb{FALSE}).  The covariates are typically
binary (0 = no, 1 = yes) responses to questions which may be predictive of
the test result, but any numeric or factor covariates can be used.}

\item{link}{the character-valued name of the link function for binomial
regression.  Choices are \verb{"logit"} (default), \verb{"cloglog"} or
\verb{"probit"}.}

\item{Nfolds}{number of folds used for \emph{k}-fold cross
validation (default = 40).}

\item{...}{additional arguments passsed to or from other \code{stats::glm}
or \code{pROC::roc}.}
}
\value{
An object of class binomscreenr containing the elements:
\describe{
\item{\code{Call}}{The function call.}
\item{\code{ModelFit}}{An object of class "glm" (See \code{\link{glm}}) containing the results of the model fit.}
\item{\code{Prevalence}}{Prevalence (proportion) of the test condition in the training sample.}
\item{\code{ParamEst}}{A vector containing the binomial regression parameter estimates.}
\item{\code{ISroc}}{A list of class "roc" (see \code{\link[pROC]{roc}}) containing in-sample (overly optimistic) results.}
\item{\code{CVpreds}}{A data frame containing \emph{k}-fold cross-validation results.}
\item{\code{CVroc}}{A list of class "roc" (See \code{\link[pROC]{roc}}) containing cross-validated results.}
}
}
\description{
Estimate binomial model parameters and cross-validated performance of test
screening based on binomial regression.  The results provide information from
which to choose a probability threshold above which individual out-of-sample
probabilies indicate the need to perform a diagnostic test.  Out-of-sample
performance is estimated using \emph{k}-fold cross validation.
}
\examples{
## Evaluate the performance of screening thresholds based on a logisitc model

data(unicorns)
help(unicorns)
unitool <- binomialScreening(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5,
                             data = unicorns, link = "logit", Nfolds = 20)
summary(unitool)
plot(unitool)
\dontrun{testCounts(unitool)}

## Example implementation of screening based on those results
## Suppose there are new observations (excluding testing) from two previously
## untested unicorns:

new <- data.frame(ID = c('"Bernie P."', '"Alice D."'), Q1 = c(0, 0), Q2 = c(0, 0),
                   Q3 = c(1, 0), Q4 = c(0, 0), Q5 = c(1, 0))
print(new)

## Compute point estimates of their predicted probabilities testing positive:
inverseLink(as.matrix(cbind(rep(1, nrow(new)), new[, 2:6])) \%*\%
                           as.matrix(unitool$ParamEst, ncol = 1), "logit")
## or, more simply,
predict(unitool$ModelFit, newdata = new, type = "response")

## If p.threshold = 0.025 is chosen as the screening threshold
## (sensitivity and specificity 77\\% and 69\\%, respectively) then "Bernie P."
## would be offered testing and "Alice D." would not.

## In practice, the computation of the probabilities of positive test results
## among newly observed individuals might be coded outside of R using, say, a
## spreadsheet.

}
\seealso{
\code{\link[stats]{glm}}

\code{\link{mebinomScreening}}
}
