% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-methods.R
\name{coef.logreg_screenr}
\alias{coef.logreg_screenr}
\title{An S3 Method to Extract Coefficients from \code{logreg_screenr} Objects}
\usage{
\method{coef}{logreg_screenr}(object, ..., intercept = TRUE, or = FALSE)
}
\arguments{
\item{object}{an object of class \code{logreg_screenr}.}

\item{...}{optional arguments passed to \code{predict} methods.}

\item{intercept}{(logical) retain (\code{TRUE}, default) or drop
(\code{FALSE}) the intercept coefficients.}

\item{or}{return odds ratios if \verb{TRUE}; logit-scale coefficients
are the default.}
}
\value{
A numeric vector containing the estimated coefficients on the logit
scale.
}
\description{
\code{coef.logreg_screenr} extracts the logistic model parameter estimates
from \code{logreg_screenr}-class objects.
}
\details{
\code{coef.logreg_screenr} extracts the estimated coefficients from
\code{logreg_screenr} objects.
}
\examples{
attach(uniobj2)
class(uniobj2)
coef(uniobj2)

}
