% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera_method.R
\name{compute_camera}
\alias{compute_camera}
\title{Compute Camera}
\usage{
compute_camera(
  xglm,
  lrt,
  DGEList,
  matrix_model,
  contrast,
  number_barcode = 3,
  thresh = 1e-04,
  lfc = 1
)
}
\arguments{
\item{xglm}{object created with \code{\link[edgeR]{estimateDisp}}}

\item{lrt}{object created with \code{\link[edgeR]{glmFit}}}

\item{DGEList}{object of edgeR}

\item{matrix_model}{The matrix that will be used to perform the
linear model analysis. Created using
\code{\link[stats]{model.matrix}}}

\item{contrast}{A vector or a single value indicating the index or the name
of the column of model_matrix to which perform the analysis}

\item{number_barcode}{Number of barcode to be considered a hit}

\item{thresh}{The threshold used to select the hits}

\item{lfc}{The Log2FC threshold}
}
\value{
The list of hits found by the camera method
}
\description{
This function computes the actual hits using camera
}
\keyword{internal}
