% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_data_table.R
\name{compute_data_table}
\alias{compute_data_table}
\title{Create the ScreenR Object}
\usage{
compute_data_table(screenR_Object)
}
\arguments{
\item{screenR_Object}{The ScreenR object}
}
\value{
screenR_Object
}
\description{
Initial function to create the Screen Object.
}
\examples{
\dontrun{
compute_data_table(object)
}

}
