% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds.R
\name{compute_explained_variance}
\alias{compute_explained_variance}
\title{Compute explained variance}
\usage{
compute_explained_variance(screenR_Object)
}
\arguments{
\item{screenR_Object}{The Object of the package}
}
\value{
A data.frame containing all the information of the variance
        expressed by the components
}
\description{
This  is an internal function  used to compute
             the explained variance by each of the Principal Components.
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))

compute_explained_variance(object)
}
\keyword{internal}
