% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute Metrics}
\usage{
compute_metrics(screenR_Object, control, treatment, day)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}

\item{control}{The control Samples.}

\item{treatment}{The treatment Samples}

\item{day}{The day of the treatment}
}
\value{
Return a tibble  with all the measure computed.
}
\description{
This function computes the  metrics that will be then used
             to compute the Z-score using the function
             \code{\link{find_zscore_hit}} starting from the screenR object
             for a given treatment in a given day
}
\examples{
obj <- get0('obj', envir = asNamespace('ScreenR'))

compute_metrics(obj, control = 'Met', treatment = 'DMSO', day = 'Day3')
}
