% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute metrics}
\usage{
compute_metrics(screenR_Object)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}.}
}
\value{
return a tibble  with all the mesures computed
}
\description{
Compute the metrics of the data_table.
             The metrix computed are:
             - Log2FC of Treated vs Control.
             - Z score of the Log2FC.
             - Robust Z score.
}
