% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute Metrics}
\usage{
compute_metrics(screenR_Object, control, treatment, day)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}

\item{control}{The vector specifying the control samples}

\item{treatment}{The vector specifying the treated samples}

\item{day}{A string containing the day to consider}
}
\value{
Return a tibble with all the measure computed.
}
\description{
This function computes the  metrics that will be then used
             to compute the Z-score using the function
             \code{\link{find_zscore_hit}} starting from the screenR object
             for a given treatment in a given day. More information can
             be found at this paper
             \href{https://pubmed.ncbi.nlm.nih.gov/21515799/}{Z-score}
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
metrics <- compute_metrics(object,
    control = "TRT",
    treatment = "Time3", day = "Time3"
)
head(metrics)

}
\concept{compute}
