% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_slope.R
\name{compute_slope}
\alias{compute_slope}
\title{Compute Slope for Gene}
\usage{
compute_slope(screenR_Object, genes, group_var)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
create_screenR_object}

\item{genes}{The genes for which the slope as to be computed. Those genes are
the result of the three statistical methods selection}

\item{group_var}{The variable to use as X for the linear model}
}
\value{
A tibble containing in each row the gene and the corresponding Slope
}
\description{
This function is used to compute the slope of the genes passed
             as input
}
