% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mapped_reads.R
\name{count_mapped_reads}
\alias{count_mapped_reads}
\title{Count the number of mapped read}
\usage{
count_mapped_reads(screenR_Object)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}
}
\value{
Return a tibble containing the number of mapped read for sample
}
\description{
This function counts the number of reads for Barcode
             in each sample
}
\examples{
obj <- get0('obj', envir = asNamespace('ScreenR'))
count_mapped_reads(obj)

}
