% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountTable_THP1_CONTROL_vs_MET.R
\docType{data}
\name{CountTable_THP1_CONTROL_vs_MET}
\alias{CountTable_THP1_CONTROL_vs_MET}
\title{Table od the count table}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{T0}{It contains an ID that identify each barcode, it can be use
  to marge the annotation table with the count table}
  \item{T48_postPURO}{It contains the cDNA sequence of the Barcode}
  \item{Day3_Met_A}{It contains the library from which the Barcode come frome}
}
}
\usage{
CountTable_THP1_CONTROL_vs_MET
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
