% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_object.R
\name{create_edger_obj}
\alias{create_edger_obj}
\title{Create edgeR Object}
\usage{
create_edger_obj(screenR_Object)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}
}
\value{
The edgeR object will all the needed information for the analysis.
}
\description{
Utility function that using the screenr-class
             object create the corresponding edgeR object.
             This function and other utility function enables the user to
             not worry abut the implementation and just focus
             on the analysis. The ScreenR package will take care of the rest.
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
create_edger_obj(object)
}
\concept{objects}
