% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edgeR_obj.R
\name{create_edgeR_obj}
\alias{create_edgeR_obj}
\title{Create EdgeR Object}
\usage{
create_edgeR_obj(screenR_Object)
}
\arguments{
\item{screenR_Object}{The Object of the package}
}
\value{
The edgeR object edgeR
}
\description{
Utility function that using the screenR-class
             object create the corresponding edgeR object
}
\examples{
obj <- get0("obj", envir = asNamespace("ScreenR"))
create_edgeR_obj(obj)
}
