% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_screenR_object.R
\name{create_screenR_object}
\alias{create_screenR_object}
\title{Create the ScreenR Object}
\usage{
create_screenR_object(
  table = NULL,
  annotation = NULL,
  groups = NULL,
  replicates = NULL
)
}
\arguments{
\item{groups}{A factor containing the experimental design label}

\item{replicates}{A vector containing the replicates label}

\item{count_table}{The count table obtained from the read alignment that
contains the Barcodes as rows and samples as columns.}

\item{annotation_table}{The annotation table containing the information
for each Barcode and the association to the
corresponding Gene}
}
\value{
An object containing all the needed information for the analysis.
}
\description{
Initial function to create the Screen Object.
}
