% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_mapped_reads.R
\name{distribution_mapped_reads}
\alias{distribution_mapped_reads}
\title{Plot the distribution of the mapped reads}
\usage{
distribution_mapped_reads(
  screenR_Object,
  palette = NULL,
  alpha = 1,
  type = "boxplot"
)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}.}

\item{palette}{The color vector for the plot.}

\item{alpha}{A value for the opacity of the plot.
Allowed values are in the range 0 to 1}

\item{type}{The type of plot. The default is "boxplot" the other option is
"density."}
}
\value{
Return a tibble containing the number of mapped read for sample
}
\description{
This function creates a boxplot to show the distribution of
             the mapped reads in different samples.
}
\examples{
object <- get0('object', envir = asNamespace('ScreenR'))

# Boxplot
distribution_mapped_reads(object)

# Density
distribution_mapped_reads(object, type = 'density')

distribution_mapped_reads(object, type = 'density', alpha = 0.2)
}
\concept{plot}
