% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_mapped_reads.R
\name{distribution_mapped_reads}
\alias{distribution_mapped_reads}
\title{Plot the distribution of the mapped reads}
\usage{
distribution_mapped_reads(
  screenR_Object,
  palette = NULL,
  alpha = 1,
  type = "boxplot"
)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}.}

\item{palette}{The color vector for the plot.}

\item{alpha}{The opacity of the plot.}

\item{type}{The type of plot. THe default is boxplot the other option is
density.}
}
\value{
return a tibble containing the number of mapped read for sample
}
\description{
Create a boxplot to show the distribution of the mapped reads
             in different sample
}
\examples{
obj <- get0('obj', envir = asNamespace('ScreenR'))

distribution_mapped_reads(obj)

distribution_mapped_reads(obj, type = 'density')

distribution_mapped_reads(obj, type = 'density', alpha = 0.5)
}
\concept{plot}
