% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by.R
\name{filter_by_slope}
\alias{filter_by_slope}
\title{Filter using the slope filter}
\usage{
filter_by_slope(
  screenR_Object,
  genes,
  group_var_treatment,
  group_var_control,
  slope_control = NULL,
  slope_treatment
)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
create_screenR_object}

\item{genes}{The genes for which the slope as to be computed. Those genes are
the result of the three statistical methods selection}

\item{group_var_treatment}{The variable to use as X for the linear model
for the Treatment}

\item{group_var_control}{The variable to use as X for the linear model
for the the control}

\item{slope_control}{The control slope}

\item{slope_treatment}{The treatment slope}
}
\value{
A data frame with the slope for the treatment and the control
        for each gene
}
\description{
This function is used to improve the quality of the hits.
             It computes a regression line in the different samples ad uses
             the slope of this line to see the trend
}
\examples{
obj <- get0("obj", envir = asNamespace("ScreenR"))

filter_by_slope(screenR_Object = obj, genes = c("SEPT5", "GLS"),
                group_var_treatment = c("T0", "T48", "Met"),
                group_var_control = c("T0", "T48", "Day3", "Day6"),
                slope_control = 0.5, slope_treatment = 1)
}
