% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera_method.R
\name{find_camera_hit}
\alias{find_camera_hit}
\title{Find Camera Hit}
\usage{
find_camera_hit(
  screenR_Object,
  matrix_model,
  contrast,
  number_barcode = 3,
  thresh = 1e-04,
  lfc = 1,
  direction = "Down"
)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}

\item{matrix_model}{The matrix that will be used to perform the
linear model analysis created using
\code{\link[stats]{model.matrix}}}

\item{contrast}{A vector or a single value indicating the index or the name
of the column the model_matrix wo which perform the analysis}

\item{number_barcode}{Number of barcode to use}

\item{thresh}{The threshold}

\item{lfc}{The Log2FC}

\item{direction}{String containing the direction of the variation}
}
\value{
The hit find with the camera method
}
\description{
This function finds the hits using the camera method is
             a wrapper for the \code{\link[limma]{camera}} function
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))

matrix <- model.matrix(~ slot(object, "groups"))
colnames(matrix) <- c("Control", "T1/T2", "Treated")

result <- find_camera_hit(
    screenR_Object = object,
    matrix_model = matrix, contrast = "Treated"
)
head(result)
}
