% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_roast_hit.R
\name{find_roast_hit}
\alias{find_roast_hit}
\title{Find Roast Hit}
\usage{
find_roast_hit(
  screenR_Object,
  matrix_model,
  contrast,
  nrot = 9999,
  number_barcode = 3,
  direction = "Down",
  p_val = 0.05
)
}
\arguments{
\item{screenR_Object}{The Object of the package}

\item{matrix_model}{the matrix that will be used to perform the
linear model analysis}

\item{contrast}{A vector or a single value indicating the index or the name
of the column the model_matrix wo which perform the analysis}

\item{nrot}{Number of rotation to perform the test}

\item{number_barcode}{The hit find with the Roast method}

\item{direction}{Direction of variation}

\item{p_val}{P val cut off}
}
\value{
The hit found by roast
}
\description{
Find the hit using the Roast method
}
\examples{
obj <- get0("obj", envir = asNamespace("ScreenR"))
matrix_model <- model.matrix(~slot(obj, "groups"))
colnames(matrix_model) <- c("Control", "T0_T48", "Treated")

find_roast_hit(obj, matrix_model = matrix_model, contrast = "Treated")
}
