% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_robust_zscore_hit.R
\name{find_robust_zscore_hit}
\alias{find_robust_zscore_hit}
\title{Title Find robust Z-score Hit}
\usage{
find_robust_zscore_hit(table_treate_vs_control, number_barcode)
}
\arguments{
\item{table_treate_vs_control}{A table coomputed with the function
\code{compute_data_table}}

\item{number_barcode}{Number of barcode to have under the median}
}
\value{
Return a tibble containing the hit for the robust Z-score
}
\description{
Title Find robust Z-score Hit
}
\examples{
obj <- get0('obj', envir = asNamespace('ScreenR'))
table <- compute_metrics(obj, control = 'Met', treatment = 'DMSO',
                         day = 'Day3')
find_robust_zscore_hit(table, number_barcode = 6)

}
