% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geebinomScreening.R
\name{geebinomScreening}
\alias{geebinomScreening}
\title{Test-Screening Tool Based on Marginal Estimation Using Generalized
Estimating Equations (GEE)}
\usage{
geebinomScreening(
  formula,
  id = NULL,
  corstr = "exchangeable",
  data = NULL,
  Nfolds = 40L,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}} defining the
testing outcome and predictor covariates, which is passed to
\code{geepack::geeglm()}.}

\item{id}{a character string naming the variable representing the sampling
clusters.}

\item{corstr}{a character string specifying the random-effect correlation
structure; one of "exchangeable", "independence", "ar1", "unstructured" or
"userdefined".  See \code{\link[geepack{geeglm}]}.}

\item{data}{the "training" sample; a data frame containing the testing
outcome and predictive covariates to be used for testing screening.  The
testing outcome must be binary (0,1) indicating negative and positive test
results, respectively, or logical (\verb{TRUE}/\verb{FALSE}).  The covariates
are typically binary (0 = no, 1 = yes) responses to questions which may be
predictive of the test result, but any numeric or factor covariates can be
used.}

\item{Nfolds}{an integer number of folds used for \emph{k}-fold cross
validation (default = 40).}

\item{...}{additional arguments passsed to \code{geepack::geeglm}.}
}
\value{
An object of class binomscreenr containing the elements:
\describe{
\item{\code{Call}}{The function call.}
\item{\code{ModelFit}}{An object of class \code{\link[geepack]{geeglm}}}
also inhereting from classes \code{gee}, \code{\link[stats]{glm}} and
\code{lm}.
\item{\code{Prevalence}}{Prevalence of the test condition in the training
sample.}
\item{\code{ParamEst}}{A vector containing the GEE parameter
estimates.}
\item{\code{ISroc}}{An object of class \code{\link[pROC]{roc}} containing
the "in-sample" (overly-optimistic) receiver operating characteristics.}
\item{\code{CVpreds}}{An object of class \code{cv.predictions} containing
the data and cross-validated predicted condition \code{y}.}
\item{\code{CVroc}}{An object of class \code{\link[pROC]{roc}} containing
the \emph{k}-fold cross-validated "out-of-sample" receiver operating
characteristics.}
}
}
\description{
Estimation of the marginal (population averaged) logistic model parameters in
the presence of cluster-level random effects, and cross-validated performance
of test screening.  The results
provide information from which to choose a probability threshold above which
individual out-of-sample probabilies indicate the need to perform a diagnostic
test.  Out-of-sample performance is estimated using \emph{k}-fold cross
validation.
}
\examples{
## Evaluate the performance of screening thresholds based on a mixed-effect
## logisitc model

##data(unicorns)
##help(unicorns)
##unitool <- geebinomScreening(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5,
##id = "clinic", data = unicorns)

##summary(unitool)
##plot(unitool)
##\dontrun{testCounts(unitool)}

## Example implementation of screening based on those results
## Suppose there are new observations (excluding testing) from two previously
## untested unicorns:

##new <- data.frame(ID = c('"Bernie P."', '"Alice D."'), Q1 = c(0, 0),
##Q2 = c(0, 0), Q3 = c(1, 0), Q4 = c(0, 0), Q5 = c(1, 0),
##clinic = factor(c("C-5", "C-15"),
##                         levels = c("C-1", "C-10", "C-11", "C-12", "C-13",
##                                    "C-14", "C-15", "C-16", "C-17,", "C-18",
##                                    "C-19", "C-2", "C-20", "C-3", "C-4",
##                                    "C-5", "C-6", "C-7", "C-8", "C-9")))
##print(new)
## Compute point estimates of their predicted probabilities testing positive:
##inverseLink("logit",
##            as.matrix(cbind(rep(1, nrow(new)), new[, 2:6])) \%*\%
##                            as.matrix(unitool$ParamEst, ncol = 1))
## or, more simply,
##predict(unitool$ModelFit, newdata = new, type = "response")
## If, for example, \code{p} = 0.025 is chosen as the screening threshold
## (sensitivity and specificity 77\\% and 69\\%, respectively) then "Bernie P."
## would be offered testing and "Alice D." would not.

## In practice, the computation of the probabilities of positive test results
## among newly observed individuals might be coded outside of R using, say, a
## spreadsheet.
}
\references{
Halekoh et al. The R package geepack for generalized
estimating equations.  Journal of Statistical Software. 2006; 15(2):1-11
\url{https://www.jstatsoft.org/article/view/v015i02}
}
\seealso{
\code{\link[geepack]{geeglm}}
}
