% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenr-class.R
\name{get_data_table}
\alias{get_data_table}
\title{Get ScreenR data_table}
\usage{
get_data_table(object)
}
\arguments{
\item{object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}
}
\value{
The data_table of the ScreenR object
}
\description{
Get function for the data_table of the ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
data_table <- get_data_table(object)
data_table
}
\concept{objects}
