% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenr-class.R
\name{get_replicates}
\alias{get_replicates}
\title{Get ScreenR replicates}
\usage{
get_replicates(object)
}
\arguments{
\item{object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}
}
\value{
The replicates of the ScreenR object
}
\description{
Get function for the replicates of the ScreenR object
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
replicates <- get_replicates(object)
replicates
}
\concept{objects}
