% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_what-methods.R
\name{get_what.easy_tool}
\alias{get_what.easy_tool}
\title{An S3 Method for Extraction of Components from \code{easy_tool} Objects}
\usage{
\method{get_what}{easy_tool}(
  from = NULL,
  what = NULL,
  ...,
  bootreps = 2000,
  conf.level = 0.95,
  se.min = 0.7
)
}
\arguments{
\item{from}{the \code{easy_tool}-class object from which to extract
the component.}

\item{what}{the (character) name of the component to extract. Valid values are
\verb{"Call"}, \verb{"QuestionWeights"}, \verb{"ROCci"}, \verb{"ROC"} and
\verb{"Scores"}.}

\item{...}{optional arguments to \code{get_what} methods.}

\item{bootreps}{the number of bootstrap replications for estimation of
confidence intervals for \code{what =} \verb{"ROCci"}.}

\item{conf.level}{(optional) confidence level for \code{what =} \verb{ROCci}}

\item{se.min}{minimum value of sensitivity printed for
\code{what =} \verb{ROCci} (default = 0.7).}
}
\value{
The selected component is returned invisibly.
}
\description{
\code{get_what.easy_tool} extracts components from \code{easy_tool}-class
objects.
}
\details{
\code{get_what} is provided to enable easy extraction of components that
are not provided by the \code{plot}, \code{predict}, \code{print} or
\code{summary} methods.

The following values of \code{what} return:
\describe{
\item{\verb{"Call"}}{the function call that created \code{from}.}
\item{\verb{"QuestionWeights"}}{the screening question weights, which
are the re-scaled logistic-regression coefficients.}
\item{\verb{ROCci}}{a data frame containing sensitivities,
specificities and their confidence limits, and thresholds}
\item{\verb{"Scores"}}{the screening scores for each subject, which are the
sums of the products of the binary question responses and their
\code{QuestionWeights}}
\item{\verb{"ROC"}}{the receiver-operating characteristic for the \code{Scores}}
}
}
\examples{
\dontrun{
attach(uniobj1)
tool <- easy_tool(uniobj1, max = 3, crossval = TRUE)
## Get and print sensitivities and specificities at thresholds for the
##   local maxima of the ROC curve
ROCci <- get_what(from = tool, what = "ROCci")
print(ROCci)
}
}
