% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_what.R
\name{get_what}
\alias{get_what}
\title{\code{get_what} is an S3 generic function to extract components of objects
produced by functions in the \code{screenr} package}
\usage{
get_what(from, what, ...)
}
\arguments{
\item{from}{an object from which to extract \code{what}.}

\item{what}{the element to extract from \code{from}.}

\item{...}{additional arguments.}
}
\description{
\code{get_what} is an S3 generic function to extract components of objects
produced by functions in the \code{screenr} package
}
\seealso{
\code{link[screenr]{get_what.lasso_screenr}}

\code{link[screenr]{get_what.logreg_screenr}}

\code{link[screenr]{get_what.simple_screenr}}
}
