% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_what-methods.R
\name{get_what}
\alias{get_what}
\title{S3 Methods for Extraction of Object Components}
\usage{
get_what(from, what, ...)
}
\arguments{
\item{from}{an object from which to extract \code{what}.}

\item{what}{the element to extract from \code{from}.}

\item{...}{additional arguments.}
}
\description{
\code{get_what} extracts components from objects.
}
\seealso{
\code{link{get_what.lasso_screenr}}

\code{link{get_what.logreg_screenr}}

\code{link{get_what.simple_screenr}}
}
