% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_what-methods.R
\name{get_what.simple_screenr}
\alias{get_what.simple_screenr}
\title{An S3 Method for Extraction of Components from \code{simple_screenr} Objects}
\usage{
\method{get_what}{simple_screenr}(
  from = NULL,
  what = NULL,
  ...,
  conf.level = 0.95,
  bootreps = 2000,
  se.min = 0.6
)
}
\arguments{
\item{from}{the \code{simple_screenr}-class object from which to extract
the component.}

\item{what}{the (character) name of the component to extract. Valid values
are \verb{"ROCci"} and \verb{"isROC"}.}

\item{...}{optional arguments to \code{get_what} methods.}

\item{conf.level}{(optional) confidence level for \code{what =} \verb{"ROCci"}}

\item{bootreps}{the number of bootstrap replications for estimation of
confidence intervals for \code{what =} \verb{"ROCci"}.}

\item{se.min}{minimum value of sensitivity printed for
\code{what =} \verb{ROCci} (default = 0.6).}
}
\value{
The selected component is returned invisibly.
}
\description{
\code{get_what.simple_screenr} extracts components from \code{simple_screenr}-class
objects.
}
\details{
\code{get_what} is provided to enable easy extraction of components for those
who wish to perform computations that are not provided by the
\code{plot}, \code{predict}, \code{print} or \code{summary} methods.

The following values of \code{what} return:
\describe{
\item{\verb{"isROC"}}{the \code{roc}-class object produced by \code{\link[pROC]{roc}}
containing the in-sample (overly optimistic) receiver-operating characteristic.}
\item{\verb{"ROCci"}}{a data frame containing cross-validated sensitivities,
specificities and their confidence limits, and thresholds.}
}
}
\examples{
\dontrun{
data(unicorns)
too_simple <- simple_screenr(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5 + Q6 + Q7,
                            data = unicorns)
too_simple_roc <- get_what(from = too_simple, what = "isROC" )
plot(too_simple_roc)
}

}
