% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getROC}
\alias{getROC}
\title{Extract ROCs from code{binomscreenr} or \code{simplescreenr} Objects}
\usage{
getROC(x, simplify = TRUE)
}
\arguments{
\item{x}{an object of class \code{binomscreenr} or \code{simplescreenr}.}

\item{simplify}{logical: simplify result to the maximum values of specificity
corresponding to unique values of sensitivity (default is \code{TRUE}).}
}
\value{
A data frame containing threshold scores, sensitivities and
specificities. Sensitivities and specificities are displayed as proportions
rather than percentages.
}
\description{
Extract the receiver operating characteristics from an object of class
\code{binomscreenr} or \code{simplescreenr}.  This is a convenience function
to enable easy use and export of the ROC.
}
\examples{
data(unicorns)
unitool <- binomialScreening(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5,
                             data = unicorns, Nfolds = 20)
(uniROC <- getROC(unitool))
}
\references{
Fawcett T. An introduction to ROC analysis. Pattern Recognition Letters. 2006.
27(8):861-874.
\url{https://doi.org/10.1016/j.patrec.2005.10.010}

Linden A. Measuring diagnostic and predictive accuracy in disease
management: an introduction to receiver operating characteristic (ROC) analysis.
Journal of Evaluation in Clinical Practice. 2006; 12(2):132-139.
\url{https://onlinelibrary.wiley.com/doi/epdf/10.1111/j.1365-2753.2005.00598.x}

Robin X, Turck N, Hainard A, Tiberti N, Lisacek F, Sanchez J-C, Muller M.
pROC: an open-source package for R and S+ to analyze and compare ROC curves.
BMC Bioinformatics 2011; 12:77. \url{https://www.biomedcentral.com/1471-2105/12/77}
}
