% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWhat.R
\name{getWhat}
\alias{getWhat}
\title{\code{getWhat} is an S3 generic function to extract components of objects
produced by functions in the \code{screenr} package}
\usage{
getWhat(from, what, ...)
}
\arguments{
\item{from}{an object from which to extract \code{what}.}

\item{what}{the element to extract from \code{from}.}

\item{...}{additional arguments.}
}
\description{
\code{getWhat} is an S3 generic function to extract components of objects
produced by functions in the \code{screenr} package
}
\seealso{
\code{link[screenr]{getWhat.glmpathScreenr}}

\code{link[screenr]{getWhat.logisticScreenr}}

\code{link[screenr]{getWhat.simpleScreenr}}
}
