% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{inverseLink}
\alias{inverseLink}
\title{Compute Inverses of Binomial Regression Link Functions}
\usage{
inverseLink(link, lp)
}
\arguments{
\item{link}{Character link function (one of "logit", "cloglog" or "probit")}

\item{lp}{Numeric linear predictor}
}
\value{
The inverse of the link function for the linear predictor.
}
\description{
Returns the inverse of logit, cloglog and probit link functions for a linear
predictor
}
