% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{inverseLink}
\alias{inverseLink}
\title{\code{inverseLink} computes inverse of logistic regression link functions}
\usage{
inverseLink(lp = NULL, link = NULL)
}
\arguments{
\item{lp}{numeric vector containing the estimated link.}

\item{link}{(character) name of the link function (one of \verb{"logit"}, \verb{"cloglog"}
or \verb{"probit"}).}
}
\value{
A numeric vector containing the inverse of the link function for the
linear predictor.
}
\description{
Returns the inverse of logit, cloglog and probit link functions for a linear
predictor
}
\details{
\code{inverseLink} returns the inverses of logit, cloglog and probit link functions,
and is provided as a (laborious) way to compute predicted values from the \verb{ModelFit}
component of \code{logisticScreenr}-class objects.  The \code{predict} methods are
a better way to obtain predicted values.
}
\note{
\code{inverseLink} may not be included in future versions of the \code{screenr}
package.
}
\examples{
## Make predictions of probability of infection from new observations
load(bsobj2)
class(bsobj2)
new_corns <- data.frame(ID = c("Alice D.", "Bernie P."),
                        testresult = c(NA, NA), Q1 = c(0, 0), Q2 = c(0, 0),
                        Q3 = c(0, 0), Q4 = c(0, 0), Q5 = c(0, 1), Q6 = c(0, 1 ))
mfit <- getWhat(what = "ModelFit", from = bsobj2 )
coefs <- mfit$coefficients
lp <- as.matrix(cbind(rep(1, nrow(new_corns)), new_corns[, 3:8])) \%*\%
           as.matrix(coefs, ncol =  1)
(preds <- inverseLink(lp, link = "logit"))
## Note that only the predicted values are returned.
}
\seealso{
\code{\link[screenr]{predict.logisticScreenr}}
}
