% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{keepfirst}
\alias{keepfirst}
\title{Return Data Frame Rows Having Unique Values in Selected Columns}
\usage{
keepfirst(x, colnames, data = NULL)
}
\arguments{
\item{x}{character-valued column name along which the dataframe is sorted.}

\item{colnames}{a character vector of column names  to identify uniqueness.}

\item{data}{a data frame.}
}
\value{
A data frame consisting of the rows of \code{data} which are
unique with respect to \code{colnames}
}
\description{
Sort a dataframe and then retain those rows which are unique with respect
to the values of selected columns.
}
