% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg_screenr.R
\name{logreg_screenr}
\alias{logreg_screenr}
\alias{{binomialScreenr}}
\title{Fitting Screening Tools Using Ordinary Logistic Regression}
\usage{
logreg_screenr(
  formula,
  data = NULL,
  link = "logit",
  Nfolds = 10,
  seed = Sys.time(),
  ...
)
}
\arguments{
\item{formula}{an object of class \code{stats::formula} defining the
testing outcome and predictor covariates, which is passed to \code{stats::glm()}.}

\item{data}{the "training" sample; a data frame containing the testing
outcome and predictive covariates to be used for testing screening.  The
testing outcome must be binary (0,1) indicating negative and positive test
results, respectively, or logical (\verb{TRUE}/\verb{FALSE}).  The covariates
are typically
binary (0 = no, 1 = yes) responses to questions which may be predictive of
the test result, but any numeric or factor covariates can be used.}

\item{link}{the character-valued name of the link function for logistic
regression.  Choices are \verb{"logit"} (default), \verb{"cloglog"} or
\verb{"probit"}.}

\item{Nfolds}{number of folds used for \emph{k}-fold cross
validation (default = 10, minimum = 2, maximum = 100).}

\item{seed}{random-number generator seed for cross-validation data splitting.}

\item{...}{additional arguments passsed to or from other \code{stats::glm}
or \code{pROC::roc}.}
}
\value{
An object of class logreg_screenr containing the elements:
\describe{
\item{\code{Call}}{The function call.}
\item{\code{formula}}{The formula object.}
\item{\code{Prevalence}}{Prevalence (proportion) of the test condition in the
training sample.}
\item{\code{ModelFit}}{An object of class \verb{glm} (See \code{\link{glm}})
containing the results of the model fit.}
\item{\code{ISroc}}{An object of class \code{\link[pROC]{roc}} containing
the "in-sample" (overly-optimistic) receiver operating characteristics,
and additional functions for use with this object are available in the
\code{pROC} package.}
\item{\code{CVpreds}}{An object of class \code{cv.predictions} containing
the data and cross-validated predicted condition \code{y}.}
\item{\code{CVroc}}{An object of class \code{\link[pROC]{roc}} containing
the \emph{k}-fold cross-validated "out-of-sample" receiver operating
characteristics, and additional functions for use with this object are
available in the \code{pROC} package.}
\item{\code{CVcoef}}{the estimated coefficients from cross-validation}
\item{\code{X_ho}}{{the matrix of held-out predictors for each cross-validation fold}}
}
}
\description{
\code{logreg_screenr} is a convenience function which integrates ordinary
logistic regression \emph{k}-fold  cross-validation and estimation of the
receiver-operating characteristic.
}
\details{
The results provide information from
which to choose a probability threshold above which individual out-of-sample
probabilies indicate the need to perform a diagnostic test.  Out-of-sample
performance is estimated using \emph{k}-fold cross validation.

The receiver operating characteristics are computed using the \code{pROC}
package. See References and package documentation for additional details.

For a gentle but python-centric introduction to \emph{k}-fold cross-validation,
see \url{https://machinelearningmastery.com/k-fold-cross-validation/}.
}
\note{
\code{logreg_screenr} is intended mainly for comparison with \code{lasso_screenr}.
Careful manual model selection is required with \code{logreg_screenr}.
\code{lasso_screenr} is easier and should generally produce better results.
}
\examples{
data(unicorns)
help(unicorns)
uniobj2 <- logreg_screenr(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5 + Q6 + Q7,
                           data = unicorns, link = "logit", Nfolds = 10)
summary(uniobj2)

}
\references{
Robin X, Turck N, Hainard A, Tiberti N, Lisacek F, Sanchez J-C,
Müller M. \code{pROC}: An open-source package for \code{R} and S+ to
analyze and compare ROC curves. BMC Bioinformatics. 2011;12(77):1-8.
\url{http://doi.org/10.1186/1471-2105-12-77}
}
\seealso{
\code{\link[stats]{glm}}
}
