% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapped_reads.R
\name{mapped_reads}
\alias{mapped_reads}
\title{Mapped Reads}
\usage{
mapped_reads(screenR_Object)
}
\arguments{
\item{screenR_object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}
}
\value{
return a tibble containing the number of mapped read for sample
}
\description{
This function perform plot the number of reads mapped for each
             sample.
}
