% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mebinomScreening.R
\name{mebinomScreening}
\alias{mebinomScreening}
\title{Test-Screening Tool Based on Marginal Estimation from Mixed-Effects Binomial
Models}
\usage{
mebinomScreening(
  formula,
  id = NULL,
  data = NULL,
  link = "logit",
  Nfolds = 40L,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}  defining the testing
outcome and predictor covariates, which is passed to \code{lme4::glmer()}.}

\item{id}{a vector (variable) which identifies the sampling clusters.}

\item{data}{the "training" sample; a data frame containing the testing
outcome and predictive covariates to be used for testing screening.  The
testing outcome must be binary (0,1) indicating negative and positive test
results, respectively, or logical (\verb{TRUE}/\verb{FALSE}).  The covariates are typically
binary (0 = no, 1 = yes) responses to questions which may be predictive of
the test result, but any numeric or factor covariates can be used.}

\item{link}{the character-valued name of the link function for binomial
regression.  Choices are \code{"logit"} (default), \code{"cloglog"} or
\code{"probit"}.}

\item{Nfolds}{an integer number of folds used for \emph{k}-fold cross
validation (default = 40).}

\item{...}{additional arguments passsed to \code{lme4::glmer}.}
}
\value{
An object of class binomscreenr containing the elements:
\describe{
\item{\code{Call}}{The function call.}
\item{\code{ModelFit}}{An object of class \code{\link[lme4]{merMod-class}}}
\item{\code{Prevalence}}{Prevalence of the test condition in the training sample.}
\item{\code{ParmEst}}{A vector containing the binomial regression parameter estimates.}
\item{\code{InSamplePerf}}{A data frame containing in-sample (overly-optimistic)
sensitivities and specificities.}
\item{\code{CrossVal}}{A data frame containing \emph{k}-fold cross-validation results.}
\item{\code{CrossValPerf}}{A data frame containing out-of-sample  sensitivities and
specificities.}
}
}
\description{
Estimation of the marginal (population averaged) binomial model parameters in
the presence of cluster-level random effects, and cross-validated performance
of test screening.  The results
provide information from which to choose a probability threshold above which
individual out-of-sample probabilies indicate the need to perform a diagnostic
test.  Out-of-sample performance is estimated using \emph{k}-fold cross
validation.
}
\examples{
## Evaluate the performance of screening thresholds based on a mixed-effect
## logisitc model

data(unicorns)
help(unicorns)
unitool <- mebinomScreening(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5, id = "clinic",
                             data = unicorns, link = "logit")
summary(unitool)
plot(unitool)
\dontrun{testCounts(unitool)}

## Example implementation of screening based on those results
## Suppose there are new observations (excluding testing) from two previously
## untested unicorns:

new <- data.frame(ID = c('"Bernie P."', '"Alice D."'), Q1 = c(0, 0), Q2 = c(0, 0),
                   Q3 = c(1, 0), Q4 = c(0, 0), Q5 = c(1, 0),
                   clinic = factor(c("C-5", "C-15"),
                                   levels = c("C-1", "C-10", "C-11", "C-12", "C-13",
                                              "C-14", "C-15", "C-16", "C-17,", "C-18",
                                              "C-19", "C-2", "C-20", "C-3", "C-4",
                                              "C-5", "C-6", "C-7", "C-8", "C-9")))
print(new)
## Compute point estimates of their predicted probabilities testing positive:
inverseLink("logit",
            as.matrix(cbind(rep(1, nrow(new)), new[, 2:6])) \%*\%
                            as.matrix(unitool$ParamEst, ncol = 1))
## or, more simply,
predict(unitool$ModelFit, newdata = new, type = "response")
## If, for example, \code{p} = 0.025 is chosen as the screening threshold
## (sensitivity and specificity 77\\% and 69\\%, respectively) then "Bernie P."
## would be offered testing and "Alice D." would not.

## In practice, the computation of the probabilities of positive test results
## among newly observed individuals might be coded outside of R using, say, a
## spreadsheet.
}
\seealso{
\code{\link[lme4]{glmer}}
}
