% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{nnt_}
\alias{nnt_}
\title{Compute the Ratio of Total Tests Performed  Per Postive Result}
\usage{
nnt_(dframe)
}
\arguments{
\item{dframe}{a data frame containing columns \code{sensitivity},
\code{specificity} and \code{prev}.}
}
\description{
\code{nnt_} computes the anticipated average number of tests
performed in order to observe a positive test result.
}
