% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{nnt_}
\alias{nnt_}
\title{\code{nnt_} computes the anticpated number of tests per postive from a
structured dataframe}
\usage{
nnt_(dframe)
}
\arguments{
\item{dframe}{a data frame containing columns \code{sensitivity},
\code{specificity} and \code{prev}.}
}
\description{
\code{nnt_} computes the anticpated number of tests per postive from a
structured dataframe
}
