% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntpp-methods.R
\name{ntpp.default}
\alias{ntpp.default}
\title{Compute the Ratio of Total Tests to Positive Results from a Data Frame}
\usage{
\method{ntpp}{default}(se = NULL, sp = NULL, prev = NULL, ...)
}
\arguments{
\item{se}{a numeric vector of sensitivities in (0,1)}

\item{sp}{a numeric vector of sensitivities in (0,1)}

\item{prev}{a numeric vector of prevalences of the testing condition, in (0,1)}

\item{...}{optional arguments to \code{ntpp} methods.}
}
\value{
a data frame containing the following columns:
\describe{
\item{\code{sensitivity}}{the sensitivity (proportion)}
\item{\code{specificity}}{the specificity (proportion)}
\item{\code{prev}}{prevalence proportion of the test condition}
\item{\code{ntpp}}{anticipated total tests required per positive result}
\item{\code{prev_untested}}{anticipated prevalence proportion among the untested}
}
}
\description{
\code{ntpp.data.frame} computes the ratio of the total number of
tests performed per positive test result from data frames.
}
\details{
The anticipated number of tests required to detect a single positive
\emph{nntp} is given by
\deqn{nntp = (SeP + (1 - Sp)(1 - P)) / SeP}
where \emph{Se} is sensitivity, \emph{P} is prevalence and \emph{Sp} is
specificity. The anticipated prevalence among those screened out is given by
\deqn{Puntested = ((1 - Se)P) / ((1 - Se)P + Sp (1 - P))}
}
