% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntpp-methods.R
\name{ntpp}
\alias{ntpp}
\title{An S3 Method to Compute the Ratio of Total Tests to Positive Results}
\usage{
ntpp(object, ...)
}
\arguments{
\item{object}{an object from which to compute the number of tests
per test positive test results.}

\item{...}{additional arguments.}
}
\value{
\code{ntpp} returns a dataframe containing the following columns:
\describe{
\item{\verb{sensitivity}}{The sensitivity (proportion) of the screener.}
\item{\verb{specificity}}{The specificity (proportion) of the screener.}
\item{\verb{ntpp}}{the number of tests required to discover
a single positive test result.}
\item{\verb{prev_untested}}{The antipated proportion who would test positive
among those who are screened out of testing.}
}
}
\description{
\code{ntpp} computes the ratio of the total number of tests
performed per positive test result and the anticipated proportion of the
untested (those screened out of testing) who would actually test positive.
}
\details{
The anticipated number of tests required to detect a single positive
\emph{nntp} is given by
\deqn{nntp = (SeP + (1 - Sp)(1 - P)) / SeP}
where \emph{Se} is sensitivity, \emph{P} is test positivity and \emph{Sp} is
specificity. The anticipated true positivity among those screened out is given by
\deqn{Puntested = ((1 - Se)P) / ((1 - Se)P + Sp (1 - P))}
}
\examples{
attach(uniobj2)
ntpp(uniobj2)

}
\seealso{
\code{\link[screenr]{ntpp.default}}, \code{\link[screenr]{ntpp.data.frame}},
\code{\link[screenr]{ntpp.easy_tool}},
\code{\link[screenr]{ntpp.lasso_screenr}},
\code{\link[screenr]{ntpp.logreg_screenr}},
\code{\link[screenr]{ntpp.simple_screenr}}
}
