% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntpp-methods.R
\name{ntpp}
\alias{ntpp}
\title{An S3 Method to Compute the Ratio of Total Tests to Positive Results}
\usage{
ntpp(object, ...)
}
\arguments{
\item{object}{an object from which to compute the number of tests
per test positive test results.}

\item{...}{additional arguments.}
}
\description{
An S3 Method to Compute the Ratio of Total Tests to Positive Results
}
\seealso{
\code{link[screenr]{ntpp.lasso_screenr}}

\code{link[screenr]{ntpp.logreg_screenr}}

\code{link[screenr]{ntpp.data.frame}}

\code{link[screenr]{ntpp.simple_screenr}}
}
