% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntpp.R
\name{ntpp.simple_screenr}
\alias{ntpp.simple_screenr}
\title{\code{ntpp.simple_screenr} is a method for computation of the anticipated
number of tests per positive test result}
\usage{
\method{ntpp}{simple_screenr}(object, ..., prev = NULL)
}
\arguments{
\item{object}{a \code{simple_screenr}-class object produced by \code{simple_screenr}.}

\item{...}{optional arguments to \code{ntpp} methods.}

\item{prev}{an optional prevalence proportion for the test outcome; if missing
the prevalence is obtained from \code{object}.}
}
\value{
A data frame containing the following columns:
\describe{
\item{\verb{sensitivity}}{The sensitivity (proportion) of the screener.}
\item{\verb{specificity}}{The specificity (proportion) of the screener.}
\item{\verb{ntpp}}{the number of tests required to discover
a single positive test result.}
\item{\verb{prev_untested}}{The prevalence proportion of the test
condition among those who are screened out of testing.}
}
}
\description{
\code{ntpp.simple_screenr} is a method for computation of the anticipated
number of tests per positive test result
}
\details{
The anticipated number of tests needed to observe a single positive test
result is a function of sensitivity, specificity and the prevalence proportion
of the condition being tested.
}
\examples{
data(unicorns)
toosimple <- simple_screenr(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5 + Q6 + Q7,
                           data = unicorns)
ntpp(toosimple)

}
