% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialScreening.R
\name{plot.binomscreenr}
\alias{plot.binomscreenr}
\title{Plot ROC Curves of \code{binomscreenr} Objects}
\usage{
\method{plot}{binomscreenr}(
  x,
  plot_ci = TRUE,
  print_ci = TRUE,
  conf_level = 0.95,
  bootreps = 2000,
  ...
)
}
\arguments{
\item{x}{An object of class "binomscreenr".}

\item{plot_ci}{Logical indicator for plotting point-wise confidence
intervals at the locally maximum subset of coordinates for
on sensitivity and specificity (default = TRUE). See also
\code{\link[pROC]{ci.thresholds}}.}

\item{print_ci}{Logical indicator to return a dataframe of numerical values,
intervals.}

\item{conf_level}{Confidence level in the interval (0,1). Default is 0.95
producing 95\% confidence intervals}

\item{bootreps}{Number of bootstrap replications for estimation of confidence
(default = 2000).}

\item{...}{additional arguments passed to \code{\link[pROC]{plot.roc}} and friends.}
}
\value{
This function produces a plot as a side effect and (optionally)
returns a dataframe dataframe containing numerical values.
}
\description{
Plot cross-validated (out-of-sample) ROC curve with pointwise confidence
intevals, along with the overly optimistic in-sample ROC curve.
}
\references{
Fawcett T. An introduction to ROC analysis. Pattern Recognition Letters. 2006.
27(8):861-874.
\url{https://doi.org/10.1016/j.patrec.2005.10.010}

Linden A. Measuring diagnostic and predictive accuracy in disease
management: an introduction to receiver operating characteristic (ROC) analysis.
Journal of Evaluation in Clinical Practice. 2006; 12(2):132-139.
\url{https://onlinelibrary.wiley.com/doi/epdf/10.1111/j.1365-2753.2005.00598.x}

Robin X, Turck N, Hainard A, Tiberti N, Lisacek F, Sanchez J-C, Muller M.
pROC: an open-source package for R and S+ to analyze and compare ROC curves.
BMC Bioinformatics 2011; 12:77. \url{https://www.biomedcentral.com/1471-2105/12/77}
}
