% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.lasso_screenr}
\alias{plot.lasso_screenr}
\title{Plot ROC Curves from \code{lasso_screenr}-Class Objects}
\usage{
\method{plot}{lasso_screenr}(
  x,
  ...,
  plot_ci = TRUE,
  model = c("minAIC", "minBIC"),
  conf_level = 0.95,
  bootreps = 4000,
  print_auc = TRUE,
  partial_auc = c(0.8, 1),
  partial_auc_focus = c("sensitivity", "specificity"),
  partial_auc_correct = TRUE,
  type = "S"
)
}
\arguments{
\item{x}{an object of class \code{lasso_screenr}.}

\item{...}{any additional arguments passed to \code{pROC::plot.roc} or
\code{pROC::lines.roc}.}

\item{plot_ci}{(logical) plot confidence intervals if \verb{TRUE}.
Default: \verb{TRUE}.}

\item{model}{(character) select either the model which produced the
minimum AIC (\verb{"minAIC"}) or minimum BIC (\verb{"minBIC"}). Default:
\verb{minAIC},}

\item{conf_level}{confidence level. Default: 0.95.}

\item{bootreps}{the number of bootstrap replications for estimation of
confidence intervals. Default: 4000.}

\item{print_auc}{logical indicator for printing the area
under the ROC curve (AUC) on the plot.  Default: \verb{TRUE}.}

\item{partial_auc}{One of \verb{FALSE} or a length two numeric vector
of the form \verb{c(a, b)} where \verb{a} and \verb{b} are the endpoints
of the interval over which to compute the partial AUC (pAUC). Ignored if
\code{print_auc = FALSE}. Default: \code{c(0.8, 1)}.}

\item{partial_auc_focus}{one of \verb{"sensitivity"} or \verb{"specificity"},
indicating the measure for which the partial AUC is to be computed. Default:
\verb{"specificity"}.}

\item{partial_auc_correct}{logical indictor for transformation of the pAUC
to fall within the range from 0.5 (random guess) to 1.0 (perfect
classification). Default: \verb{TRUE}.}

\item{type}{type of plot. See \code{\link[base]{plot}}. Default: \verb{"S"}.}
}
\value{
This function produces a plot as a side effect.
}
\description{
\code{plot.lasso_screenr} plots the \emph{k}-fold cross-validated
receiver-operating characteristic for out-of-sample screening performance,
including confidence intervals on the combinations of the local maxima of
sensitivity and specificity.
}
\details{
Plot cross-validated (out-of-sample) ROC curve with pointwise confidence
intevals along with the overly optimistic in-sample ROC curve.
\code{plot.lasso_screenr} is an enhanced convenience wrapper for
\code{pROC::plot.roc}.
}
\examples{
\dontrun{
attach(uniobj1)
plot(uniobj1, model = "minAIC")
}
}
\references{
Fawcett T. An introduction to ROC analysis. Pattern Recognition Letters. 2006.
27(8):861-874.
\url{https://doi.org/10.1016/j.patrec.2005.10.010}

Linden A. Measuring diagnostic and predictive accuracy in disease
management: an introduction to receiver operating characteristic (ROC) analysis.
Journal of Evaluation in Clinical Practice. 2006; 12(2):132-139.
\url{https://onlinelibrary.wiley.com/doi/epdf/10.1111/j.1365-2753.2005.00598.x}

Robin X, Turck N, Hainard A, Tiberti N, Lisacek F, Sanchez J-C, Muller M.
pROC: an open-source package for R and S+ to analyze and compare ROC curves.
BMC Bioinformatics 2011; 12:77. \url{https://www.biomedcentral.com/1471-2105/12/77}
}
