% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.logreg_screenr}
\alias{plot.logreg_screenr}
\title{Plot ROC Curves from \code{logreg_screenr}-Class Objects}
\usage{
\method{plot}{logreg_screenr}(x, ..., plot_ci = TRUE, conf_level = 0.95, bootreps = 2000)
}
\arguments{
\item{x}{an object of class \code{logreg_screenr}.}

\item{...}{additional arguments passed to \code{\link[pROC]{plot.roc}} and friends.}

\item{plot_ci}{logical indicator for plotting point-wise confidence
intervals at the locally maximum subset of coordinates for
on sensitivity and specificity (default = \verb{TRUE}). See also
\code{\link[pROC]{ci.thresholds}}.}

\item{conf_level}{confidence level in the interval (0,1). Default is 0.95
producing 95\% confidence intervals.}

\item{bootreps}{number of bootstrap replications for estimation of confidence
(default = 2000).}
}
\value{
This function produces a plot as a side effect.
}
\description{
\code{plot.logreg_screenr} plots the \emph{k}-fold cross-validated
receiver-operating characteristic, including confidence intervals on the
combinations of the local maxima of sensitivity and specificity.
}
\details{
Plot cross-validated (out-of-sample) ROC curve with pointwise confidence
intevals along with the overly optimistic in-sample ROC curve.
\code{plot.lasso_screenr} is an enhanced convenience wrapper for
\code{pROC::plot.roc}.
}
\examples{
attach(uniobj2)
plot(uniobj2)

}
\references{
Fawcett T. An introduction to ROC analysis. Pattern Recognition Letters. 2006.
27(8):861-874.
\url{https://doi.org/10.1016/j.patrec.2005.10.010}

Linden A. Measuring diagnostic and predictive accuracy in disease
management: an introduction to receiver operating characteristic (ROC) analysis.
Journal of Evaluation in Clinical Practice. 2006; 12(2):132-139.
\url{https://onlinelibrary.wiley.com/doi/epdf/10.1111/j.1365-2753.2005.00598.x}

Robin X, Turck N, Hainard A, Tiberti N, Lisacek F, Sanchez J-C, Muller M.
pROC: an open-source package for R and S+ to analyze and compare ROC curves.
BMC Bioinformatics 2011; 12:77. \url{https://www.biomedcentral.com/1471-2105/12/77}
}
