% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScreening.R
\name{plot.simplescreenr}
\alias{plot.simplescreenr}
\title{An S3 Plot Method for \code{simplescreenr} Objects}
\usage{
\method{plot}{simplescreenr}(x)
}
\arguments{
\item{x}{A simplescreenr class object.}
}
\description{
An S3 Plot Method for \code{simplescreenr} Objects
}
