% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode_hit.R
\name{plot_barcode_hit}
\alias{plot_barcode_hit}
\title{Plot Barcode Hit}
\usage{
plot_barcode_hit(
  screenR_Object,
  matrix_model,
  contrast,
  number_barcode = 3,
  gene,
  quantile = c(-0.5, 0.5),
  labels = c("Negative logFC", "Positive logFC")
)
}
\arguments{
\item{screenR_Object}{The Object of the package}

\item{matrix_model}{the matrix that will be used to perform the
linear model analysis}

\item{contrast}{An object created with makeContrasts
function}

\item{number_barcode}{Number of barcode to have under the median}

\item{gene}{a single gene to plot}

\item{quantile}{Quantile to display on the plot}

\item{labels}{Title of the plot}
}
\value{
A vector containing the common hit
}
\description{
Create a Barcode plot for the Hit
}
\examples{
obj <- get0('obj', envir = asNamespace('ScreenR'))
matrix_model <- model.matrix(~slot(obj, 'groups'))
colnames(matrix_model) <- c('Control', 'T0_T48', 'Treated')
contrast <- limma::makeContrasts(Treated - Control, levels = matrix_model)

plot_barcode_hit(obj, matrix_model, contrast = contrast, gene = 'ACACB')
}
\concept{plot}
