% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode_hit.R
\name{plot_barcode_hit}
\alias{plot_barcode_hit}
\title{Plot Barcode Hit}
\usage{
plot_barcode_hit(
  screenR_Object,
  matrix_model,
  contrast,
  number_barcode = 3,
  gene,
  quantile = c(-0.5, 0.5),
  labels = c("Negative logFC", "Positive logFC")
)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}

\item{matrix_model}{the matrix that will be used to perform the
linear model analysis}

\item{contrast}{An object created with \code{\link[limma]{makeContrasts}}
function.}

\item{number_barcode}{Number of barcode to have under the median}

\item{gene}{A  gene name to plot}

\item{quantile}{Quantile to display on the plot}

\item{labels}{Title of the plot}
}
\value{
The barcode plot
}
\description{
Create a Barcode plot for the Hit
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))
matrix_model <- model.matrix(~ slot(object, "groups"))
colnames(matrix_model) <- c("Control", "T1_T2", "Treated")
contrast <- limma::makeContrasts(Treated - Control, levels = matrix_model)

plot_barcode_hit(object, matrix_model, contrast = contrast,
                 gene = "Gene_300")

}
\concept{plot}
