% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_lost.R
\name{plot_barcode_lost}
\alias{plot_barcode_lost}
\title{Plot number of barcode lost}
\usage{
plot_barcode_lost(
  screenR_Object,
  palette = NULL,
  alpha = 1,
  legende_position = "none"
)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}

\item{palette}{A vector of colors}

\item{alpha}{A value for the opacity of the plot}

\item{legende_position}{Where to positioning the legend of the plot}
}
\value{
return a tibble containing the number of mapped read for sample
}
\description{
This function plots the number of barcode lost in each sample
}
\concept{plot}
