% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode_lost.R
\name{plot_barcode_lost}
\alias{plot_barcode_lost}
\title{Plot number of barcode lost}
\usage{
plot_barcode_lost(screenR_Object, palette = NULL, alpha = 1)
}
\arguments{
\item{screenR_object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}
}
\value{
return a tibble containing the number of mapped read for sample
}
\description{
This fuction count the number of Barcode lost during the
             sequenceing. A barcode is lost if it has zero mapped read
}
