% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_lost.R
\name{plot_barcode_lost_for_gene}
\alias{plot_barcode_lost_for_gene}
\title{Plot number of barcode lost for gene}
\usage{
plot_barcode_lost_for_gene(screenR_Object, facet = TRUE, samples = NULL)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenr_object}}}

\item{facet}{A boolean to use the facet}

\item{samples}{The samples to visualize}
}
\value{
Return a tibble containing the number of mapped reads for sample
}
\description{
This function plots the number of barcodes lost in each sample
}
\examples{
object <- get0("object", envir = asNamespace("ScreenR"))

plot_barcode_lost_for_gene(object,
    samples = c("Time3_A", "Time3_B")
)
plot_barcode_lost_for_gene(object,
    samples = c("Time3_A", "Time3_B"),
    facet = FALSE)
}
\concept{plot}
