% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode_trend.R
\name{plot_barcode_trend}
\alias{plot_barcode_trend}
\title{Plot the trend over time of the Barcodes}
\usage{
plot_barcode_trend(
  list_data_measure,
  genes,
  n_col = 2,
  size_line = 1,
  color = NULL
)
}
\arguments{
\item{list_data_measure}{A list containing the measure table of the
different time point}

\item{genes}{Vector of genes name}

\item{n_col}{The number of column to use in the facet wrap}

\item{size_line}{Rhe dimension of the line}

<<<<<<< HEAD
\item{color}{A color palette for the plot}
=======
\item{color}{Vector of colors}
>>>>>>> develop_2
}
\value{
The rend plot for the genes in input
}
\description{
Plot the log2FC over time of the barcodes in the different
             time point
}
