% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Plot Barcode Hit}
\usage{
plot_boxplot(
  screenR_Object,
  genes,
  group_var,
  alpha = 0.5,
  nrow = 1,
  ncol = 1,
  fill_var = "Sample",
  type = "boxplot"
)
}
\arguments{
\item{screenR_Object}{The Object of the package}

\item{genes}{The vector of genes to use}

\item{group_var}{The variable that as to be used to filter the data, for
example the different treatment}

\item{alpha}{The opacity of the regression line}

\item{nrow}{The number of rows in case multiple genes are plotted}

\item{ncol}{The number of columns in case multiple genes are plotted}

\item{fill_var}{The variable used to fill the boxplot}

\item{type}{The type of plot to use boxplot or violinplot}
}
\value{
A boxplot
}
\description{
This function plot the boxplot in each sample for the genes
             passed as input
}
\examples{
obj <- get0("obj", envir = asNamespace("ScreenR"))

plot_boxplot(obj, genes = c("SEPT5", "GLS"),
             group_var = c("T0", "T48", "Met"), nrow = 1, ncol = 2,
             fill_var = "Day", type = "violinplot")
}
\concept{plot}
