% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_lost.R
\name{plot_distribution_of_barcode_lost}
\alias{plot_distribution_of_barcode_lost}
\title{Plot distribution of barcode lost}
\usage{
plot_distribution_of_barcode_lost(
  screenR_Object,
  palette = NULL,
  alpha = 1,
  type = "boxplot"
)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}

\item{palette}{A vector of colors}

\item{alpha}{A value for the opacity of the plot.
Allowed values are in the range 0 to 1}

\item{type}{Type of plot. Allowed values are "boxplot" and "boxplot"}
}
\value{
Return a tibble containing the number of mapped read for sample
}
\description{
The function plots the distribution of the lost barcodes in
             each sample
}
\examples{
#object <- get0('object', envir = asNamespace('ScreenR'))
#plot_distribution_of_barcode_lost(object)
#plot_distribution_of_barcode_lost(object, type = 'density')

}
\concept{plot}
