% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_lost.R
\name{plot_distribution_of_barcode_lost}
\alias{plot_distribution_of_barcode_lost}
\title{Plot distribution of barcode lost}
\usage{
plot_distribution_of_barcode_lost(
  screenR_Object,
  palette = NULL,
  alpha = 1,
  type = "boxplot"
)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}

\item{palette}{A vector of colors}

\item{alpha}{A value for the opacity of the plot}

\item{type}{Type of plot}
}
\value{
return a tibble containing the number of mapped read for sample
}
\description{
The function plot the distribution of the lost barcode in each
             sample
}
