% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapped_reads.R
\name{plot_mapped_reads}
\alias{plot_mapped_reads}
\title{Plot Mapped Reads}
\usage{
plot_mapped_reads(
  screenR_Object,
  palette = NULL,
  alpha = 1,
  legende_position = "none"
)
}
\arguments{
\item{screenR_Object}{The ScreenR object obtained using the
\code{\link{create_screenR_object}}}

\item{palette}{A vector of color}

\item{alpha}{The alpha parameter for the opacity of the plot}
}
\value{
return a ggplot object
}
\description{
This function perform plot the number of reads mapped for each
             sample.
}
