% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MDS.R
\name{plot_MDS}
\alias{plot_MDS}
\title{Multidimensional Scaling Plot}
\usage{
plot_MDS(
  screenR_Object,
  groups = NULL,
  alpha = 0.8,
  size = 2.5,
  color = "black"
)
}
\arguments{
\item{screenR_Object}{The Object of the package}

\item{groups}{The vector to fill the plot}

\item{alpha}{The opacity of the labels}

\item{size}{The dimension of the labels}

\item{color}{The color of the labels}
}
\value{
The MDS Plot
}
\description{
Plot samples on a two-dimensional scatterplot so that
             distances on the plot approximate the typical log2 fold
             changes between the samples.
}
\examples{
obj <- get0('obj', envir = asNamespace('ScreenR'))

plot_MDS(obj)
}
