% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MDS.R
\name{plot_MDS}
\alias{plot_MDS}
\title{PLot the Sample in vector Space}
\usage{
plot_MDS(screenR_Object, palette, dimension = 2, groups = NULL)
}
\arguments{
\item{screenR_Object}{The Object of the package}

\item{palette}{palette to use to color the groups}

\item{dimension}{number of plotting dimensions}

\item{groups}{variable to fill the plot}
}
\value{
The MDS Plot
}
\description{
Plot samples on a two or three-dimensional scatterplot so that
             distances on the plot approximate the typical log2 fold
             changes between the samples.
}
