% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MDS.R
\name{plot_PC_explained_variance}
\alias{plot_PC_explained_variance}
\title{PLot the explained variance by the PC}
\usage{
plot_PC_explained_variance(
  screenR_Object,
  cumulative = FALSE,
  color = "steelblue"
)
}
\arguments{
\item{screenR_Object}{The ScreenR object}

\item{cumulative}{A boolean value which indicates whether or not to plot
the cumulative variance}

\item{color}{The color to fill the barplot}
}
\value{
The explained variance  plot
}
\description{
This function plot the explained variance by the
             Principal Component.
}
\examples{
object <- get0('object', envir = asNamespace('ScreenR'))

plot_PC_explained_variance(object)

# For the cumulative plote
plot_PC_explained_variance(object, cumulative = TRUE)
}
