% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot the trend Hit Gene}
\usage{
plot_trend(
  screenR_Object,
  genes,
  group_var,
  alpha = 0.5,
  se = FALSE,
  point_size = 1,
  line_size = 1,
  nrow = 1,
  ncol = 1
)
}
\arguments{
\item{screenR_Object}{The ScreenR object}

\item{genes}{The vector of genes to use}

\item{group_var}{The variable that as to be used to filter the data, for
example the different treatment}

\item{alpha}{The opacity of the regression line}

\item{se}{A boolean to indicate where or not to plot the standard error}

\item{point_size}{The dimension of each dot}

\item{line_size}{The dimension of the line}

\item{nrow}{The number of rows in case multiple genes are plotted}

\item{ncol}{The number of columns in case multiple genes are plotted}
}
\value{
The the plot of the
}
\description{
This function plot the trend of a gene resulted as hit of the
             analysis
}
\examples{
obj <- get0('obj', envir = asNamespace('ScreenR'))

plot_trend(obj, genes = 'GLS', group_var = c('T0', 'T48', 'Met'))

plot_trend(obj, genes = c('SEPT5', 'GLS'), group_var = c('T0', 'T48', 'Met'),
   nrow = 2)

}
