% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Zscore_distribution.R
\name{plot_Zscore_distribution}
\alias{plot_Zscore_distribution}
\title{Plot distribution Z-score}
\usage{
plot_Zscore_distribution(time_point_measure, alpha = 1)
}
\arguments{
\item{time_point_measure}{A list containing the time the table for each time
point}

\item{alpha}{the alpha value for the plot}
}
\value{
return the density plot of the distribution of the Z-score
}
\description{
This function plot the distribution of the Z-score of the Log2FC of the treated
vs control in the different timepoint
}
