% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialScreening.R
\name{print.binomscreenr}
\alias{print.binomscreenr}
\title{An S3 Print Method for \code{binomscreenr} Objects.}
\usage{
\method{print}{binomscreenr}(x, quote = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{binomscreenr} produced by function.}

\item{quote}{logical, indicating whether or not strings should be printed
with surrounding quotes.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
An S3 Print Method for \code{binomscreenr} Objects.
}
