% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.easy_tool}
\alias{print.easy_tool}
\title{\code{print.easy_tool} is a print method for \code{easy_tool} objects}
\usage{
\method{print}{easy_tool}(x, ...)
}
\arguments{
\item{x}{an object of class \code{easy_tool}}

\item{...}{optional arguments to \code{print} methods.}
}
\description{
\code{print.easy_tool} is a print method for \code{easy_tool} objects
}
\examples{
attach(uniobj1)
print(uniobj1)
}
