% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.easyTool}
\alias{print.easyTool}
\title{\code{print.easyTool} is a print method for \code{easyTool} objects}
\usage{
\method{print}{easyTool}(x, ...)
}
\arguments{
\item{x}{an object of class \code{easyTool}}

\item{...}{optional arguments to \code{print} methods.}
}
\description{
\code{print.easyTool} is a print method for \code{easyTool} objects
}
\examples{
attach(uniobj1)
print(uniobj1)
}
