% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmpathScreening.R
\name{print.glmpathScreenr}
\alias{print.glmpathScreenr}
\title{code{print.glmpathScreenr} is a print method for \code{glmpathScreenr} objects}
\usage{
\method{print}{glmpathScreenr}(object)
}
\arguments{
\item{object}{an object of class \code{glmpathScreenr}}
}
\description{
code{print.glmpathScreenr} is a print method for \code{glmpathScreenr} objects
}
\examples{
attach(uniobj1)
print(uniobj1)
}
