% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lasso_screenr}
\alias{print.lasso_screenr}
\title{code{print.lasso_screenr} is a print method for \code{lasso_screenr} objects}
\usage{
\method{print}{lasso_screenr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{lasso_screenr}}

\item{...}{optional arguments to \code{print} methods.}
}
\description{
code{print.lasso_screenr} is a print method for \code{lasso_screenr} objects
}
\examples{
attach(uniobj1)
print(uniobj1)
}
