% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.logreg_screenr}
\alias{print.logreg_screenr}
\title{A Print Method for \code{logreg_screenr}-Class Objects}
\usage{
\method{print}{logreg_screenr}(x, ..., quote = FALSE)
}
\arguments{
\item{x}{an object of class \code{logreg_screenr}.}

\item{...}{optional arguments to \code{print} methods.}

\item{quote}{logical indicator for whether or not strings should be printed.}
}
\value{
Nothing. Thresholds, specificities and sensitivities are printed as a
side effect.
}
\description{
\code{print.logreg_screenr} is a print method for
\code{logreg_screenr}-class objects.
}
\examples{
attach(uniobj2)
print(uniobj2)

}
\seealso{
\code{get_what.logreg_screenr(from, what)} for \code{what =} \verb{"ROCci"}.
}
