% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.simple_screenr}
\alias{print.simple_screenr}
\title{A Print Method for \code{simple_screenr}-Class Objects}
\usage{
\method{print}{simple_screenr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{simple_screenr}.}

\item{...}{optional arguments to \code{print} methods.}
}
\value{
Nothing. Thresholds, specificities and sensitivities are printed as a
side effect.
}
\description{
\code{print.simple_screenr} is print method for
\code{simple_screenr} objects.
}
\examples{
data(unicorns)
toosimple <- simple_screenr(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5 + Q6,
                          data = unicorns)
print(toosimple)

}
\seealso{
\code{get_what.simple_screenr(from, what)} for \code{what =} \verb{"ROCci"}.
}
