% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScreening.R
\name{print.simplescreenr}
\alias{print.simplescreenr}
\title{An S3 Print Method for \code{simplescreenr} Objects}
\usage{
\method{print}{simplescreenr}(x)
}
\arguments{
\item{x}{A \code{simplescreenr} class object.}
}
\description{
An S3 Print Method for \code{simplescreenr} Objects
}
