% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{rescale_to_int}
\alias{rescale_to_int}
\title{\code{rescale_to_int} rescales all non-zero elements of a non-negative
numeric matrix or vector to integers from 1 to \verb{max}}
\usage{
rescale_to_int(x, max, colwise = TRUE)
}
\arguments{
\item{x}{numeric matrix or vector of positive real numbers.}

\item{max}{the value of largest element in the rescaled integer-valued
vector.}

\item{colwise}{(logical) rescale the matrix by column if \verb{TRUE} (the default) or
by row if \verb{FALSE}.}
}
\value{
A matrix of integers corresponding to \code{x} in which smallest
\emph{non-zero} element in each column/row is 1 and the largest element is \code{max}. Any
elements having value zero are unchanged. If \code{x} is a vector then the result is
a \emph{r} x 1 matrix, where \emph{r} is the number of elements in \code{x}.  Otherwise
the result is a \emph{r} x \emph{c} matrix where \emph{c} is the number of columns in \code{x}
}
\description{
\code{rescale_to_int} rescales all non-zero elements of a non-negative
numeric matrix or vector to integers from 1 to \verb{max}
}
\note{
Any values of 0 in \code{x} are not rescaled, and are preserved in
the result.
}
\examples{
x <- c(0.5, 1.2, 0.9, 0, 0.1)
rescale_to_int(x, max = 5)
}
\seealso{
\code{\link[scales]{rescale}}
}
