% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{roc_ci}
\alias{roc_ci}
\title{Compute Cross-Validated Sensitivities, Specificities and Uncertainty Limits}
\usage{
roc_ci(
  roc_,
  bootreps,
  conf.level,
  progress = "text",
  thresholds = "local maximas",
  se.min = se.min
)
}
\arguments{
\item{roc_}{an object of class \code{roc}.}

\item{bootreps}{number of bootstrap replicates.}

\item{conf.level}{confidence level for uncertainty intervals.}

\item{progress}{type of progress display.}

\item{thresholds}{type of thresholds.}

\item{se.min}{minimum value of sensitivity returned.}
}
\value{
a data frame containing thresholds with sensititives, specificities
and uncertainy intervals.
}
\description{
Compute Cross-Validated Sensitivities, Specificities and Uncertainty Limits
}
