% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_aov.R
\name{screen_aov}
\alias{screen_aov}
\title{screen_aov}
\usage{
screen_aov(df, responses, factors = NULL)
}
\arguments{
\item{df, }{data.frame, a \code{data.frame} with factors and responses as columns}

\item{responses, }{character, a vector of characters corresponding to column names
in \code{df} to be used as responses}

\item{factors, }{character, a vector of characters corresponding to column names
in \code{df} to be used as factors}
}
\value{
A list with the results data frame in \code{results} and inputs in \code{input}.
}
\description{
Use ANOVA to screen numeric variables to numeric factors
}
\details{
\code{screen_aov} uses \code{aov} to screen responses against factors one to one
}
\examples{
out <- screen_aov(mtcars, c("cyl", "disp"), c("wt"))

}
