% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenR-pacakage.R
\docType{package}
\name{ScreenR-package}
\alias{ScreenR-package}
\alias{ScreenR}
\title{Tools for analyzing shRNAs screening}
\description{
`ScreenR` implements a series of functions to facilitate inspection,
analysis and visualization of shRNAs screening datasets. `ScreenR`
takes the a count table as input and create the screenr_object to perform
the analysis.
}
\details{
`ScreenR` anable the user to perform quality control, visual inspection,
dimensionality reduction of the data. It uses statistical methods to find
}
\author{
Emanuel Michele Soda \email{emanuelsoda@gmail.com}
}
\keyword{internal}
