% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenr-package.R
\docType{package}
\name{screenr}
\alias{screenr}
\title{screenr Package}
\description{
The \code{screenr} package enables construction of binary test-screening
tools.  It is designed to enable those with only a basic familiarity with
\code{R} to develop, validate and implement screening tools for diagnostic
tests.  \code{screenr} integrates the capabilities of the \code{glm},
\code{glmpath} and \code{pROC} packages for convenience and ease of use.

Consider the situation where a diagnostic test for some condition is
relatively expensive, and the condition is rare.  In that case, universal
testing would not be efficient in terms of the yield of postive results per
test performed.  Now suppose that responses to a set of simple screening
questions may be predictive of the condition.  Package \code{screenr} enables
estimation of thresholds for making decisions about when to test in order to
screen in/out individuals based on Receiver Operating Characteristics (ROC)
estimated from an initial sample.  The choice of a particular screening
threshold is left to the user, and should be based on careful consideration
of application-specific tradeoffs between sensitivity and specificity.
\code{screenr} also enables easy construction of screening tools.

A tutorial is available from \code{vignette("screenr_Tutorial", package = "screenr")}.

The pdf versions of the package manual and the tutorial are available at
\url{https://github.com/sgutreuter/screenr}.
}
\details{
The high-level functions in the \code{screenr} package are:
\describe{
\item{\code{\link{lasso_screenr}}}{Selection of logistic models based on GLM
path regularization}
\item{\code{\link{logreg_screenr}}}{Test-screening based maximum-likelhood
estimation of logistic models}
\item{\code{\link{gee_screenr}}}{Test-screening based GEE estimation
of logistic models accounting for cluster sampling}
\item{\code{\link{easy_tool}}}{Easy implementation of test-screening tools}
\item{\code{\link{simple_screenr}}}{(Too) simple un-optimized test-screening}
\item{\code{\link{rescale_to_int}}}{Rescale a strictly positive vector of
real numbers to integers}
\item{\code{\link{sens_spec_plus}}}{Sensitivity, specificity and friends}
}

\code{screenr} provides the usual \code{plot}, \code{print}, \code{summary},
\code{predict} methods for the objects produced by \code{lasso_screenr},
\code{logreg_screenr}, \code{gee_screenr}, \code{simple_screenr} and
\code{easy_tool}, and also
\code{coef} and \code{confint} methods for \code{lasso_screenr},
\code{gee_screenr} and \code{logreg_screenr} objects.
\code{screenr} also provides \code{get_what} methods to extract object
components, and \code{ntpp} methods for computation of the average number of
tests required to detect a single positive result and the residual
positivity among those screened out of testing.
}
\note{
The canonical source repository for \code{screenr} is
\url{https://github.com/sgutreuter/screenr}
}
\references{
Teferi W, Gutreuter S, Bekele A et al. Adapting strategies for effective and
efficient pediatric HIV case finding: Risk screening tool for testing children
presenting at high-risk entry points. BMC Infectious Diseases. 2022; 22:480.
\url{http://doi.org/10.1186/s12879-022-07460-w}
}
\author{
Steve Gutreuter: \email{sgutreuter@gmail.com}
}
