% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenr-class.R
\docType{class}
\name{screenr_object-class}
\alias{screenr_object-class}
\alias{screenr_object}
\title{The screenr_object Class
The screenr_object class is the main object of the package, it is passed
to a series of function to perform the analysis.}
\description{
The screenr_object Class
The screenr_object class is the main object of the package, it is passed
to a series of function to perform the analysis.
}
\section{Slots}{

\describe{
\item{\code{count_table}}{It is used to store the count table to perform the analysis}

\item{\code{annotation_table}}{It is used to store the annotation of the shRNA}

\item{\code{groups}}{It is used to store the vector of treated and untreated}

\item{\code{replicates}}{It is used to store information about the replicates}

\item{\code{normalized_count_table}}{It is used to store a normalized verision of the
count table}

\item{\code{data_table}}{It is used to store a tidy format of the count table}
}}

\concept{objects}
