% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera_method.R
\name{select_number_barcode}
\alias{select_number_barcode}
\title{Select  number of Barcode}
\usage{
select_number_barcode(gene, gene_symbols, number_barcode)
}
\arguments{
\item{gene}{The gene name}

\item{...}{
  Arguments passed on to \code{\link[=unique_gene_symbols]{unique_gene_symbols}}
  \describe{
    \item{\code{gene_symbols}}{The gene symbols list}
  }}
}
\value{
The barcode of the gene passed as input
}
\description{
Compute a unique gene symbol for gene
}
\keyword{internal}
