% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScreening.R
\name{simpleScreening}
\alias{simpleScreening}
\title{Simple Un-optimized Test-Screening Tool}
\usage{
simpleScreening(formula, data)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}} defining the
testing outcome and predictor covariates.}

\item{data}{the "training" sample; a data frame containing the testing outcome
and predictive covariates to be used for testing screening.  The testing
outcome must be binary (0,1) indicating negative and positive test results,
respectively, or logical (\verb{TRUE}/\verb{FALSE}), and the screening scores
are the row-wise sums of the values of those covariates.  The covariates are
typically binary (0 = no, 1 = yes) responses to questions, but the responses
may also be ordinal numeric values.}
}
\value{
An object of class \code{simplescreenr} containing the elements:
\describe{
\item{\code{Call}}{The function call.}
\item{\code{Prevalence}}{Prevalence of the test condition in the training sample.}
\item{\code{ISroc}}{An object of class \code{\link[pROC]{roc}} containing
the "in-sample" (overly-optimistic) receiver operating characteristics.}
\item{\code{Scores}}{The training sample, including the scores.}
}
}
\description{
Compute the in-sample (\emph{overly optimistic}) performances for development
of a very simple test screening tool implementing the method described in the
reference.  The results provide information from which to
choose a counting-based threshold score above which a diagnostic test would be
performed. However, it is unlikely that the target sensitivity and specificity
would be acheived from new data.  \code{simpleScreening} is intended for use
\emph{only} in those situations where the technical capacity for implementation
of screening is limited to counting responses to questions.  The methods
implemented by \code{binomialScreening} and \code{mebinomScreening} will almost
certaily out-perform \code{simpleScreening}.
}
\examples{
data(unicorns)
simple <- simpleScreening(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5,
                          data = unicorns)
summary(simple)
plot(simple)
\dontrun{testCounts(simple)}

}
\references{
Bandason et al 2016. Validation of a screening tool to
identify older children living with HIV in primary care facilities in high
HIV prevalence settings. AIDS 30(5):779-785
\url{http://dx.doi.org/10.1097/QAD.0000000000000959}
}
\seealso{
\code{\link{binomialScreening}}, \code{\link{mebinomScreening}}
}
