% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScreening.R
\name{simpleScreening}
\alias{simpleScreening}
\title{A Simple Un-optimized Test Screening Tool}
\usage{
simpleScreening(formula, data)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}} defining the testing
outcome and predictor covariates.}

\item{data}{The "training" sample; a data frame containing the testing outcome
and predictive covariates to be used for testing screening.  The testing
outcome must be binary (0,1) indicating negative and positive test results,
respectively, or logical (TRUE/FALSE).  The covariates are typically binary
(0 = no, 1 = yes) responses to questions, but the responses may also be
ordinal numeric values.}
}
\value{
An object of class "simplescreenr" containing the elements:
\describe{
\item{\code{Call}}{The function call.}
\item{\code{Prevalence}}{Prevalence of the test condition in the training sample.}
\item{\code{InSamplePerf}}{A data frame containing in-sample (overly optimistic)
sensitivities and specificities.}
\item{\code{Scores}}{The training sample, including the scores.}
}
}
\description{
Compute the in-sample performances for development of a very simple test
screening tool. The results provide information from which to choose a
counting-based threshold score above which a diagnostic test would be
performed. \code{binomialScreening} will almost certainly outperform this
approach, and \code{simpleScreening} is intended for use in those situations
where post-estimation technical capacity is limited to counting responses to
questions.
}
\examples{
data(unicorns)
simple <- simpleScreening(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5 ,
                       data = unicorns)
summary(simple)
plot(simple)
testCounts(simple)

}
\references{
Bandason et al 2016. Validation of a screening tool to
identify older children living with HIV in primary care facilities in high
HIV prevalence settings. AIDS 30(5):779-785
\url{http://dx.doi.org/10.1097/QAD.0000000000000959}
}
