% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialScreening.R
\name{summary.binomscreenr}
\alias{summary.binomscreenr}
\title{An S3 Summary Method for \code{binomscreenr} Objects.}
\usage{
\method{summary}{binomscreenr}(object, ...)
}
\arguments{
\item{object}{an object of class \code{binomscreenr} produced by function
\code{binomialScreening}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
An S3 Summary Method for \code{binomscreenr} Objects.
}
