% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\name{summary.easy_tool}
\alias{summary.easy_tool}
\title{An S3 Summary Method for \code{screenr} Objects}
\usage{
\method{summary}{easy_tool}(object, ...)
}
\arguments{
\item{object}{an \code{easy_tool} object.}

\item{...}{optional arguments passed to \code{summary} methods.}
}
\value{
Nothing.  A summary is printed as a side effect.
}
\description{
\code{summary.easy_tool} provides a summary method for
\code{easy-tool}-class objects.
}
\examples{
attach(uniobj1)
summary(uniobj1)
}
