% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\name{summary.easy_tool}
\alias{summary.easy_tool}
\title{A Summary Method for \code{easy_tool}-Class Objects}
\usage{
\method{summary}{easy_tool}(object, ...)
}
\arguments{
\item{object}{an \code{easy_tool} object.}

\item{...}{optional arguments passed to \code{summary} methods.}
}
\value{
a dataframe containing the summary, including the Df, Deviance,
AIC and BIC for each step along the GLM path for which the active set
changed.
}
\description{
A Summary Method for \code{easy_tool}-Class Objects
}
\details{
This is essentially a wrapper for \code{glmpath::summary.glmpath}
provided for \code{lasso_screenr} objects.
}
\examples{
attach(uniobj1)
summary(uniobj1)
}
