% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticScreening.R
\name{summary.logisticScreenr}
\alias{summary.logisticScreenr}
\title{\code{summary.logisticScreenr} is an S3 print method for \code{logisticScreenr}
objects.}
\usage{
\method{summary}{logisticScreenr}(object, diagnostics = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{logisticScreenr} produced by function
\code{logisticScreenr}.}

\item{diagnostics}{a logical value; plot model diagnostics if \verb{TRUE}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Nothing.  Summaries are printed as a side effect.
}
\description{
\code{summary.logisticScreenr} is an S3 print method for \code{logisticScreenr}
objects.
}
\examples{
load(uniobj2)
class(uniobj2)
summary(uniobj2)
}
