% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\name{summary.logreg_screenr}
\alias{summary.logreg_screenr}
\title{A Summary Method for \code{logreg_screenr}-Class Objects}
\usage{
\method{summary}{logreg_screenr}(object, ..., diagnostics = FALSE)
}
\arguments{
\item{object}{an object of class \code{logreg_screenr} produced by function
\code{logreg_screenr}.}

\item{...}{optional arguments passed to \code{summary} methods.}

\item{diagnostics}{a logical value; plot model diagnostics if \verb{TRUE}.}
}
\value{
Nothing.  Summaries are printed as a side effect.
}
\description{
\code{summary.logreg_screenr} provides a summary method for
\code{logreg_screenr}-class objects.
}
\examples{
attach(uniobj2)
summary(uniobj2)
}
