% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScreening.R
\name{summary.simplescreenr}
\alias{summary.simplescreenr}
\title{An S3 Summary Method for \code{simplescreenr} Objects}
\usage{
\method{summary}{simplescreenr}(x)
}
\arguments{
\item{x}{A simplescreenr class object.}
}
\description{
An S3 Summary Method for \code{simplescreenr} Objects
}
