% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScreening.R
\name{summary.simplescreenr}
\alias{summary.simplescreenr}
\title{Print Summaries of \code{simplescreenr} Objects}
\usage{
\method{summary}{simplescreenr}(object, ...)
}
\arguments{
\item{object}{A \code{simplescreenr} class object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Nothing.  Thresholds, specificities and sensitivities are printed as
a side effect.
}
\description{
Print Summaries of \code{simplescreenr} Objects
}
\seealso{
\code{\link{getROC}}
}
