% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScreening.R
\name{summary.simplescreenr}
\alias{summary.simplescreenr}
\title{An S3 Summary Method for \code{simplescreenr} Objects}
\usage{
\method{summary}{simplescreenr}(object, ...)
}
\arguments{
\item{object}{A \code{simplescreenr} class object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
An S3 Summary Method for \code{simplescreenr} Objects
}
