% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{testCounts}
\alias{testCounts}
\title{Expected Number of Tests Required per Positive Test Result}
\usage{
testCounts(x = NULL, prev = NULL)
}
\arguments{
\item{x}{A data frame containing columns "sensitivity" and
'"specificity"', or an object of class 'simplescreenr' or 'binomscreenr'.}

\item{prev}{Numeric proportion of the population expressing positive test
results.  \code{prev} is optional for class 'simplescreenr' and
'binomscreenr' objects, and defaults to prevalence in the training sample
if not specified.}
}
\value{
A data frame containing sensitivity, specificity, the expected
number of tests required to observe a single positive test result and,
among those, the expected number of false negatives per positive test
result.
}
\description{
Compute the expected number of tests which need to be performed in order
to identify the first positive test result, and the expected number of
false positives among that number of tests.
}
\examples{
data(unicorns)
unitool <- binomialScreening(testresult ~ Q1 + Q2 + Q3 + Q4 + Q5,
                             data = unicorns, Nfolds = 20)
testCounts(unitool)

}
