% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniobj1-data.R
\docType{data}
\name{uniobj1}
\alias{uniobj1}
\title{A \code{glmpathScreenr object}}
\format{
An object of class \code{glmpathScreenr}
}
\usage{
class(uniobj1)
}
\description{
The result of \code{uniobj1 <- glmpathScreenr(testresult ~ Q1 + Q2 + Q3 +
Q4 + Q5 + Q6, data = unicorns, Nfolds = 10, seed = 123)}
}
\examples{
\dontrun{
summary(uniobj1)
}
}
\keyword{glmpathScreenr}
