% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera_method.R
\name{unique_gene_symbols}
\alias{unique_gene_symbols}
\title{Unique gene Symbols}
\usage{
unique_gene_symbols(gene_symbols, number_barcode = 3)
}
\arguments{
\item{gene_symbols}{The gene symbols list}

\item{number_barcode}{The number of barcode to select}
}
\value{
A list of unique gene symbols
}
\description{
Compute a unique gene symbol for gene
}
